% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bariatric}
\alias{bariatric}
\title{Bariatric surgery metabolic profiling data}
\format{
A list with three elements:
\describe{
\item{X.pqn}{Matrix of pre-processed proton NMR spectra (rows = spectra)}
\item{ppm}{Chemical shift vector (ppm) with length equal to number of columns in \code{X.pqn}}
\item{meta}{Spectrometer metadata}
}
}
\usage{
data(bariatric)
}
\description{
Metabolic profiling NMR dataset of murine urine samples collected before and after bariatric surgery (Roux-en-Y gastric bypass).
}
\details{
One dimensional proton NMR spectra of murine urine samples were acquired on a Bruker Avance III spectrometer (600.13 MHz, 300 K). A standard NMR pulse sequence (\emph{recycle delay - 90° - t1 - 90° - tm - 90° - acquisition}) was applied, where \emph{t1} was 3 μs and \emph{tm} (mixing time) was 100 ms. Water suppression was achieved using selective irradiation during a recycle delay of 2 s and \emph{tm}. A 90° pulse of 10 μs was used. A total of 128 scans were collected into 64k data points with a spectral width of 20 ppm.

For further information on sample collection and processing see Li et al. (2011).
}
\references{
Li, Jia V., et al. (2011) Metabolic surgery profoundly influences gut microbial-host metabolic cross-talk. \emph{Gut}. 60(9), 1214-1223. \url{https://www.ncbi.nlm.nih.gov/pubmed/21572120}
}
\keyword{datasets}
