% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bline.R
\name{bline}
\alias{bline}
\alias{blcor}
\title{Baseline correction for 1D NMR spectra}
\usage{
bline(X, lambda = 1e+07, iter_max = 30)
}
\arguments{
\item{X}{num matrix or data.frame, NMR data with rows representing spectra}

\item{lambda, }{num smoothing paramter}

\item{iter_max}{num, maximum number of iterations}
}
\value{
Baseline corrected NMR matrix an with same dimension as input matrix.
}
\description{
Non-linear baseline correction for NMR spectra based on asymetric least squares. This function estimates smooth and non-linear baseline trends for each spectrum. The baseline is then subtracted from the respective spectrum and the remainder is returned. For more infor on smoothing parameter \emph{lambda}, check out \code{\link[ptw]{asysm}}.
}
\examples{
data(covid_raw)
X1_bc=bline(X[1,])
spec(X[1,], ppm, shift=c(3,4))
spec(X1_bc[1,], ppm, shift=c(3,4), add=TRUE)
}
\seealso{
\code{\link[ptw]{asysm}}

Other NMR: 
\code{\link{alignSegment}()},
\code{\link{excise1d}()},
\code{\link{get.idx}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{normqref}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{spec}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{stocsy}()},
\code{\link{storm}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR}
