% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1d.R
\name{checkFiles1d}
\alias{checkFiles1d}
\title{Check for intact file systems - helper function read1d}
\usage{
checkFiles1d(datapath, procs_exp = 1, n_max = 10, filter = T)
}
\arguments{
\item{datapath}{char, File directory containing spectra}

\item{procs_exp}{num, Topspin processing experiment ID}

\item{n_max}{int, Maximum number of spectra to read-in}

\item{filter}{lobic, filter for intact file systems (TRUE is recommended)}
}
\description{
Check for intact file systems - helper function read1d
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
