% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1d.R
\name{chem_shift}
\alias{chem_shift}
\title{Calc chemical shift axis points - helper function read1d}
\usage{
chem_shift(swidth, offset, si)
}
\arguments{
\item{swidth}{num, sweep width extracted from procs file. See fct extract_pars1d}
}
\description{
Calc chemical shift axis points - helper function read1d
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
