% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid}
\alias{covid}
\title{COVID-19 blood plasma proton NMR spectra (processed)}
\format{
An object of class \code{list} of length 3.
}
\usage{
covid
}
\description{
1D NMR spectra from SARS-CoV-2 positive patients (n=10) and healthy controls (n=13), collected in Perth, Western Australia.
Spectra were pre-processed including excision of residual water and signal-free areas, baseline correction, and normalization to account for different line widths.
FIDs were acquired using a standard 90˚ RF pulse sequence on a 600 MHz Avance II spectrometer using Bruker IVDr methods for blood plasma (300 K, 32 scans).
The spectrometer was equipped with a double resonance broadband probe (BBI) and a refrigerated autosampler o
}
\keyword{datasets}
