% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.calibTsp}
\alias{.calibTsp}
\title{Calibrate}
\usage{
.calibTsp(spec, ppm)
}
\value{
Shift-adjusted ppm vector
}
\description{
Calibrate
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
\keyword{internal}
