% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawGenerics.R
\name{.check1d_files_fid}
\alias{.check1d_files_fid}
\title{Check for intact file systems - helper function read1d}
\usage{
.check1d_files_fid(datapath, n_max = 10, filter = TRUE, recursive, verbose)
}
\arguments{
\item{datapath}{char, File directory containing spectra}

\item{n_max}{int, Maximum number of spectra to read-in}

\item{filter}{lobic, filter for intact file systems (TRUE is recommended)}

\item{procs_exp}{num, Topspin processing experiment ID}
}
\value{
List of strings, each one is a path to fid and experiment number
}
\description{
Check for intact file systems - helper function read1d
}
\section{}{
NA
}

\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\keyword{internal}
