% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.check_pc_model}
\alias{.check_pc_model}
\title{Check Validity of Selected Principal or Orthogonal Component}
\usage{
.check_pc_model(pc, mod, le = 1, type = "p")
}
\arguments{
\item{pc}{Character or numeric. Selected component, e.g., \code{1} or \code{"o1"}.}

\item{mod}{An object of class \code{PCA_metabom8} or \code{OPLS_metabom8}.}

\item{le}{Integer. Expected maximum length of \code{pc}; usually 1.}

\item{type}{Character. Type of component to check: \code{"p"} for projection or \code{"t"} for scores. Currently only used for logic branching.}
}
\value{
Nothing. Function is used for its side effect (i.e., error checking).
}
\description{
Internal consistency check for principal (or orthogonal) component selection used in plotting or extraction functions.
}
\keyword{internal}
