% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.checkXclassNas}
\alias{.checkXclassNas}
\title{Check X matrix (PLS context)}
\usage{
.checkXclassNas(X)
}
\arguments{
\item{X}{Input data (uni or multivar) formatted as vector or matrix or data.frame}
}
\value{
NULL if all's fine and throws error otherwise
}
\description{
This function checks for missing values in Y (NA, NAN, infinite), numeric variable format.
}
\section{}{
NA
}

\keyword{internal}
