% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.dimX}
\alias{.dimX}
\title{Transform to row vec}
\usage{
.dimX(X)
}
\arguments{
\item{X}{Array or matrix - NMR data with spectra represented in rows.}
}
\description{
Transform to row vec
}
\section{}{
NA
}

\keyword{internal}
