% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_allviz.R
\name{.draw_base}
\alias{.draw_base}
\title{Draw NMR Spectra Using Base Graphics}
\usage{
.draw_base(dat, add, ...)
}
\arguments{
\item{dat}{List returned by \code{.prep_spec()}.}

\item{add}{Logical. Add to existing plot.}

\item{...}{Additional graphical parameters.}
}
\description{
Internal backend renderer for base R plotting.
}
\keyword{internal}
