% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_allviz.R
\name{.draw_ggplot2}
\alias{.draw_ggplot2}
\title{Draw NMR Spectra Using ggplot2}
\usage{
.draw_ggplot2(dat, ...)
}
\arguments{
\item{dat}{List returned by \code{.prep_spec()}.}

\item{...}{Additional arguments for \code{ggplot2}.}
}
\description{
Internal backend renderer using \code{ggplot2}. Converts spectra
to long format prior to plotting.
}
\keyword{internal}
