% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_allviz.R
\name{.draw_plotly}
\alias{.draw_plotly}
\title{Draw NMR Spectra Using Plotly}
\usage{
.draw_plotly(dat, col = NULL, ...)
}
\arguments{
\item{dat}{List returned by \code{.prep_spec()}.}

\item{...}{Additional arguments passed to \code{plotly}.}
}
\description{
Internal backend renderer for interactive plotting.
}
\keyword{internal}
