% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.extMeanCvFeat}
\alias{.extMeanCvFeat}
\title{Extract data structures from output of function .oplsComponentCv}
\usage{
.extMeanCvFeat(cv_obj, feat = "t_xp")
}
\arguments{
\item{cv_obj}{named list, output of .oplsComponentCv (for resp OPLS component)}

\item{feat}{char, feature name in output of tt (names(tt))}
}
\value{
num array or matrix of resp component feature
}
\description{
Extract data structures from output of function .oplsComponentCv
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
