% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawGenerics.R
\name{.extract_acq_pars1d}
\alias{.extract_acq_pars1d}
\title{Read Bruker NMR paramter files - helper function read1d}
\usage{
.extract_acq_pars1d(f_list)
}
\arguments{
\item{f_list}{list, intact files system for NMR experiments. See fct checkFiles1d}

\item{procs_exp}{num or char, which processing experiment should be extracted}
}
\value{
data frame of spectrometer acquisition parameters
}
\description{
Read Bruker NMR paramter files - helper function read1d
}
\section{NA}{

}

\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\keyword{internal}
