% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1d.R
\name{.extract_pars1d}
\alias{.extract_pars1d}
\title{Read Bruker NMR Parameter Files}
\usage{
.extract_pars1d(f_list)
}
\arguments{
\item{f_list}{List of file paths. Output from \code{.detect1d_procs()}.}
}
\value{
A data frame of extracted acquisition and processing metadata. Row names correspond to spectrum filenames.
}
\description{
Helper function used by \code{read1d()} to extract acquisition (\code{acqus}) and processing (\code{procs}) parameters
from Bruker-formatted 1D NMR experiments.
}
\keyword{internal}
