% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.kFold}
\alias{.kFold}
\title{k-fold cross-validation index generator}
\usage{
.kFold(k, Y)
}
\arguments{
\item{k}{Integer. Number of CV folds. If \code{k} is not valid or too high relative to number of rows in Y,
it defaults to leave-one-out CV (LOO-CV).}

\item{Y}{Matrix. Outcome matrix (observations × variables).}
}
\value{
A list of length \code{k}, where each element contains the training indices (integers)
for one CV fold.
}
\description{
Creates a list of training set indices for k-fold cross-validation (CV),
without considering class balance in Y. Each element of the list represents
the row indices used for training in one CV fold.
}
\examples{
Y <- matrix(rnorm(100), nrow = 20)  # Simulated outcome matrix
folds <- .kFold(k = 5, Y = Y)
length(folds)         # Should be 5
sapply(folds, length) # Training set size per fold
}
\keyword{internal}
