% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.kFoldStratified}
\alias{.kFoldStratified}
\title{Y-stratified k-fold subsets for cross validation (CV)}
\usage{
.kFoldStratified(k, stratified)
}
\arguments{
\item{k}{int k parameter}

\item{stratified}{list with 3 elements: 1) type char: R for regression (Y is numeric) or DA for discriminant analysis (Y treated as categorical), 2) Y matrix (observations times variables) and 3) quantile function probabilities for stratification of numberic Y}
}
\value{
list of Y row-indices for each fold
}
\description{
This function ccreates a list of indices that represent k fold subsets for CV, stratified by the first column in Y
}
\section{}{
NA
}

