% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.mcBalanced}
\alias{.mcBalanced}
\title{k-fold Monte Carlo Cross Validation (MCCV) subsets}
\usage{
.mcBalanced(k, split, stratified)
}
\arguments{
\item{k}{int k parameter}

\item{split}{Fraction of samples used to generate training set for class/group-balanced Monte Carlo (MC) CV}

\item{stratified}{list with 3 elements: 1) type char: R for regression (Y is numeric) or DA for discriminant analysis (Y treated as categorical), 2) Y matrix (observations times variables) and 3) quantile function probabilities for stratification of numberic Y}
}
\value{
list of Y row-indices for each fold
}
\description{
This function creates a list of indices that represent k fold subsets for MCCV
}
\section{}{
NA
}

