% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.orthModelCompSummary}
\alias{.orthModelCompSummary}
\title{Summarise model indices using tabular and graphic modes}
\usage{
.orthModelCompSummary(type, r2x_comp, r2_comp, q2_comp, aucs, cv)

.orthModelCompSummary(type, r2x_comp, r2_comp, q2_comp, aucs, cv)
}
\arguments{
\item{type}{char, OPLS type: DA (discriminant analysis) or R (regression)}

\item{r2x_comp}{num array, r2x value for each component}

\item{r2_comp}{num array, r2y value for each component}

\item{q2_comp}{num array, q2 value for each component}

\item{aucs}{num array, aucs value for each component (defined only for DA)}

\item{cv}{list of cross-validion paramters (see opls function)}
}
\value{
List of two: 1. summary data.frame and 2. ggplot2 figure

list: 1. data.fram, summary table 2. ggplot2, viz of summary tbl
}
\description{
Summarise model indices using tabular and graphic modes

Summarise component indices
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}

Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
