% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.r2}
\alias{.r2}
\title{R2 and Q2}
\usage{
.r2(Y, Yhat, ytss)
}
\arguments{
\item{Y}{num matrix Y: OPLS outcome, dummy matrix in case it is categorical}

\item{Yhat}{num matrix, predicted Y}

\item{ytss}{num, total sum of squares of Y}
}
\value{
R2=1-(PRESS/TSS)
}
\description{
R2 and Q2
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
\keyword{internal}
