% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.scaleMatRcpp}
\alias{.scaleMatRcpp}
\title{Column-wise matrix scaling  (C++ backend)}
\usage{
.scaleMatRcpp(X, idc, center, scale_type)
}
\arguments{
\item{X}{num matrix}

\item{idc}{int row indices of X}

\item{center}{logical mean centering}

\item{scale_type}{int 0: no scaling, 1: SD scaling, 2: Pareto scaling}
}
\value{
list: 1. scale X matrix, 2. mean (sd vec), 3: sd (num vec)
}
\description{
Column-wise matrix scaling  (C++ backend)
}
\examples{
X <- matrix(rnorm(20), ncol = 4)
idc <- seq_len(nrow(X)) -1 # ! cpp indexing starts at zero !
.scaleMatRcpp(X, idc, TRUE, 1L)
}
\keyword{internal}
