% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.sdRcpp}
\alias{.sdRcpp}
\title{Column-wise standard deviation and mean for a matrix}
\usage{
.sdRcpp(X)
}
\arguments{
\item{X}{num matrix}
}
\value{
list: 1. sd (num vec), 2. mean (sd vec)
}
\description{
Column-wise standard deviation and mean for a matrix
}
\keyword{internal}
