% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.viz_df_helper}
\alias{.viz_df_helper}
\title{Prep data.frame for plotting functions}
\usage{
.viz_df_helper(obj, pc, an, type = "p")
}
\description{
Prep data.frame for plotting functions
}
\section{}{
NA
}

\keyword{internal}
