% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-s4_generics.R, R/03-s4_methods.R
\name{fitted}
\alias{fitted}
\alias{fitted,m8_model-method}
\title{Extract fitted Y values}
\usage{
fitted(object, ...)

\S4method{fitted}{m8_model}(object)
}
\arguments{
\item{object}{An object of class \code{m8_model}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Numeric vector or matrix containing the fitted response values.
}
\description{
Extract fitted Y values
}
\examples{
data(covid)
cv <- balanced_mc(k=5, split=2/3)
scaling <- UVScaling(center=TRUE)
model <-opls(X=covid$X, Y=covid$an$type, scaling, cv)
show(model)
Y_hat_dummy <- fitted(model)
}
