% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{get.idx}
\alias{get.idx}
\title{Find indices in ppm vector for respective chemical shift range}
\usage{
get.idx(range = c(1, 5), ppm)
}
\arguments{
\item{range}{num, range chemical shift (in ppm vector)}

\item{ppm}{num, ppm vector}
}
\description{
Find indices in ppm vector for respective chemical shift range
}
\author{
Torben Kimhofer \email{tkimhofer@gmail.com}
}
