% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hiit_raw}
\alias{hiit_raw}
\title{High-intensity interval training (HIIT) 1H NMR urine dataset}
\format{
A list with four elements:
\describe{
\item{X}{Numeric matrix of spectral intensities (samples x variables).}
\item{ppm}{Numeric vector of chemical shift values corresponding to columns of \code{X}.}
\item{meta}{Data frame containing acquisition metadata.}
\item{df}{Data frame containing sample annotations.}
}
}
\source{
Example dataset bundled with the package.
}
\description{
Urine samples collected from a single individual performing a \eqn{VO_2max}-type
exercise protocol over a time period of 3h.
}
\details{
The spectra were acquired on a 600 MHz Bruker NMR spectrometer.
The dataset is included for demonstration of preprocessing and
modelling workflows in \pkg{metabom8}.
}
\examples{
data(hiit_raw)

}
\keyword{datasets}
