% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_provenance.R
\name{list_preprocessing}
\alias{list_preprocessing}
\title{List available preprocessing steps}
\usage{
list_preprocessing()
}
\value{
A named character vector with preprocessing function
identifiers as names and short descriptions as values.
}
\description{
Returns a named character vector describing the preprocessing
operations implemented in \pkg{metabom8}. For more information
on individual functionalities please refer to the function help
pages.
}
\examples{
list_preprocessing()
}
