% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-s4_methods.R
\name{loadings,m8_model-method}
\alias{loadings,m8_model-method}
\title{Model loadings}
\usage{
\S4method{loadings}{m8_model}(x, orth = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{m8_model}.}

\item{orth}{Logical indicating whether orthogonal scores should be returned
(only applicable for OPLS models).}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Numeric vector or matrix containing loadings.
}
\description{
Model loadings
}
\examples{
data(covid)
cv <- balanced_mc(k=5, split=2/3)
scaling <- UVScaling(center=TRUE)
model <-opls(X=covid$X, Y=covid$an$type, scaling, cv)
show(model)
P <- loadings(model)
Po <- loadings(model, orth = TRUE)
dim(P)
dim(Po) == dim(P)
}
