% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-s4_classes.R, R/03-s4_methods.R
\docType{class}
\name{m8_model-class}
\alias{m8_model-class}
\alias{m8_model}
\alias{summary,m8_model-method}
\alias{show,m8_model-method}
\title{m8_model class
Model object returned by \code{pca()}, \code{pls()}, and \code{opls()}.}
\usage{
\S4method{summary}{m8_model}(object)

\S4method{show}{m8_model}(object)
}
\arguments{
\item{object}{An object of class \code{m8_model}.}
}
\value{
An object of class \code{m8_model}.
}
\description{
m8_model class
Model object returned by \code{pca()}, \code{pls()}, and \code{opls()}.
}
\section{Functions}{
\itemize{
\item \code{summary(m8_model)}: Summarise model performance and component selection.

\item \code{show(m8_model)}: Show a compact model header.

}}
\section{Slots}{

\describe{
\item{\code{engine}}{Character. Model engine ("pca", "pls", "opls").}

\item{\code{ctrl}}{List. Engine-specific control and performance information.}

\item{\code{fit}}{List. Fitted data (engine specific).}

\item{\code{cv}}{Resampling instance (may be NULL if not used).}

\item{\code{prep}}{Scaling and centering information}

\item{\code{provenance}}{Preprocessing attributes of spectral matrix X}

\item{\code{session}}{R-session information}

\item{\code{call}}{Function call}

\item{\code{dims}}{List with \code{n} and \code{p}.}
}}

\examples{
data(covid)
cv <- balanced_mc(k=5, split=2/3)
scaling <- UVScaling(center=TRUE)
model <-opls(X=covid$X, Y=covid$an$type, scaling, cv)
class(model)
show(model)
}
