% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-s4_classes.R
\name{mc}
\alias{mc}
\title{Monte-Carlo cross-validation strategy}
\usage{
mc(k, split)
}
\arguments{
\item{k}{Integer. Number of repeated random splits.}

\item{split}{Numeric. Fraction of samples assigned to the
training set (e.g. \code{2/3}).}
}
\value{
An object of class \code{MonteCarlo}.
}
\description{
Monte-Carlo cross-validation strategy
}
\details{
Monte-Carlo cross-validation (also called repeated holdout)
generates \code{k} random train/test splits without replacement.
Each split uses the proportion defined by \code{split}
for the training set.
}
\examples{
n <- 100
# bivariate outcome
thr <- 1.5
Y <- c(rnorm(80, thr-3, 0.3), rnorm(20, thr+3, 0.3))  # unbalanced low/high outcome
mean(Y>thr)

cv_mc <- mc(k = 10, split = 2/3)

mc_inst <- instantiate(cv_mc, matrix(Y, ncol = 1))
sapply(train_idc(mc_inst), function(i) length(i))
}
\seealso{
Other resampling strategies: 
\code{\link{balanced_boot}()},
\code{\link{balanced_mc}()},
\code{\link{instantiate}()},
\code{\link{kfold}()},
\code{\link{stratified_kfold}()}
}
\concept{resampling strategies}
