% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{minmax}
\alias{minmax}
\title{Min-max scaling}
\usage{
minmax(x)
}
\arguments{
\item{x}{num array to be scaled}
}
\value{
num array, same length as x
}
\description{
Min-max scaling
}
\details{
Output values are scaled to range between 0 and 1
\deqn{x_{s}=\frac{x-x_{min}}{x_{max}-x_{min}}}
}
\section{}{
NA
}

\examples{
x=rnorm(20, 0, 1)
plot(x, type='l'); abline(h=range(x), lty=2)
points(minmax(x), type='l', col='red'); abline(h=c(0,1), col='red', lty=2)
}
\seealso{
Other NMR ++: 
\code{\link{.hotellingsT2}()},
\code{\link{cvanova}()},
\code{\link{es_cdelta}()},
\code{\link{opls_perm}()},
\code{\link{plotscores}()},
\code{\link{pqn}()},
\code{\link{predict_opls}()}
}
\author{
\email{torben.kimhofer@gmail.com}
}
\concept{NMR ++}
