% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normErectic.R
\name{normErectic}
\alias{normErectic}
\title{Normalise by the ERETIC signal}
\usage{
normErectic(X, integr = FALSE)
}
\arguments{
\item{X}{num matrix, NMR data, spectra in row with ppm values as column names}

\item{integr}{logic, FALSE: fct returns integral of ERETIC, TRUE: fct returns spectra normalised with ERETIC integral}
}
\value{
\itemize{
\item num matrix: X-normalised with ERETIC integral (urine - 12 ppm, plasma = 15 ppm)
\item num vector: ERETIC integrals (urine - 12 ppm, plasma = 15 ppm)
}
}
\description{
Normalise by the ERETIC signal
}
\section{}{
NA
}

\seealso{
Other NMR: 
\code{\link{alignSegment}()},
\code{\link{bline}()},
\code{\link{excise1d}()},
\code{\link{get.idx}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{spec}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{stocsy}()},
\code{\link{storm}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR}
