% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normQref.R
\name{normqref}
\alias{normqref}
\title{Normalise with quantRef}
\usage{
normqref(X, qref = F)
}
\arguments{
\item{X}{num matrix:  NMR data, spectra in row with ppm values as column names}
}
\value{
\itemize{
\item num matrix: X-normalised with QuantRef integral (urine - 12 ppm, plasma = 15 ppm)
}
}
\description{
Normalise with quantRef
}
\section{}{
NA
}

\seealso{
Other NMR: 
\code{\link{alignSegment}()},
\code{\link{bline}()},
\code{\link{excise1d}()},
\code{\link{get.idx}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{spec}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{stocsy}()},
\code{\link{storm}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR}
