% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-s4_classes.R
\name{NoScaling}
\alias{NoScaling}
\title{No Scaling
Leaves variables unscaled. Optional centering.}
\usage{
NoScaling(center = FALSE)
}
\arguments{
\item{center}{Logical. If TRUE, variables are mean-centered before scaling.}
}
\value{
An object of class \code{m8_preprocess}.
}
\description{
No Scaling
Leaves variables unscaled. Optional centering.
}
\examples{
just_centering <- NoScaling(center=TRUE)
X <- matrix(c(10,10, 0,0, 0, 10), ncol=3)
X_centered <- prep(just_centering, X)
str(X_centered)
X_centered$X
}
\seealso{
Other scaling objects & strategies: 
\code{\link{ParetoScaling}()},
\code{\link{UVScaling}()}
}
\concept{scaling objects & strategies}
