% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opls_perm.R
\name{opls_perm}
\alias{opls_perm}
\title{OPLS model Y-permutations}
\usage{
opls_perm(smod, n = 10, plot = TRUE, mc = FALSE)
}
\arguments{
\item{smod}{OPLS model of the package \emph{metabom8}}

\item{n, }{num Number of permutations}

\item{plot}{logical, indicating if results should be visualised}

\item{mc}{logical, indicating if tasked should be parallelised using multiple cores}
}
\value{
data.frame with perutation indices
}
\description{
Model validation using Y-permutations
}
\section{}{
NA
}

\seealso{
Other OPLS model validation functions: 
\code{\link{cvanova}()},
\code{\link{dmodx}()}
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}

Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
\concept{OPLS model validation functions}
