% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{OplsMate-class}
\alias{OplsMate-class}
\title{An S4 class to represent an OPLS model constructed with MetaboMate}
\description{
An S4 class to represent an OPLS model constructed with MetaboMate
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
