% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scalings.R
\name{pareto_scaling}
\alias{pareto_scaling}
\title{Pareto Scaling
Leaves variables unscaled. Optional centering.}
\usage{
pareto_scaling(center = FALSE)
}
\arguments{
\item{center}{Logical. If TRUE, variables are mean-centered before scaling.}
}
\value{
A \code{list} with elements:
\describe{
\item{X}{Numeric matrix containing the scaled data.}
\item{prep}{List describing the preprocessing, including centering and scaling
parameters (\code{center}, \code{scale}, \code{X_mean}, \code{X_sd}).}
}
}
\description{
Pareto Scaling
Leaves variables unscaled. Optional centering.
}
\details{
Scales variables by the square root of their standard deviation.
}
\examples{
paritalUV <- pareto_scaling(center=TRUE)
X <- matrix(c(10,10, 0,0, 0, 10, 0, 1000), ncol=4)
X_scaled <- prep_X(paritalUV, X)
str(X_scaled)
X_scaled$X
}
\seealso{
Other scaling_strategies: 
\code{\link{unscaled}()},
\code{\link{uv_scaling}()}
}
\concept{scaling_strategies}
