% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-s4_classes.R
\name{ParetoScaling}
\alias{ParetoScaling}
\title{Pareto Scaling
Scales variables by the square root of their standard deviation.}
\usage{
ParetoScaling(center = FALSE)
}
\arguments{
\item{center}{Logical. If TRUE, variables are mean-centered before scaling.}
}
\value{
An object of class \code{m8_preprocess}.
}
\description{
Pareto Scaling
Scales variables by the square root of their standard deviation.
}
\details{
Pareto scaling reduces the influence of large-variance variables
while preserving data structure better than full unit variance scaling.
}
\examples{
paritalUV <-ParetoScaling(center=TRUE)
X <- matrix(c(10,10, 0,0, 0, 10, 0, 1000), ncol=4)
X_scaled <- prep(paritalUV, X)
str(X_scaled)
X_scaled$X
}
\seealso{
Other scaling objects & strategies: 
\code{\link{NoScaling}()},
\code{\link{UVScaling}()}
}
\concept{scaling objects & strategies}
