% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PCA_metabom8-class}
\alias{PCA_metabom8-class}
\title{PCA model object from metabom8}
\description{
An S4 class to represent a PCA model constructed using metabom8.
}
\section{Slots}{

\describe{
\item{\code{type}}{Character string indicating model type (e.g., "PCA")}

\item{\code{t}}{Scores matrix (PCs)}

\item{\code{p}}{Loadings matrix}

\item{\code{nPC}}{Number of principal components}

\item{\code{X_mean}}{Mean of each feature in X (numeric)}

\item{\code{X_sd}}{Standard deviation of each feature in X (numeric)}

\item{\code{Parameters}}{List of parameters used}

\item{\code{X}}{Original input matrix}
}}

\seealso{
Other NMR & MS: 
\code{\link{OPLS_metabom8-class}},
\code{\link{PLS_metabom8-class}}
}
\author{
\email{tkimhofer@gmail.com}
}
\concept{NMR & MS}
