% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{plot,OPLS_metabom8-method}
\alias{plot,OPLS_metabom8-method}
\title{Plotting function OplsMAte object}
\usage{
\S4method{plot}{OPLS_metabom8}(x, y)
}
\arguments{
\item{x}{OplsMAte object}

\item{y}{string, describing plot type: scores, scores_cv}
}
\value{
graphics output
}
\description{
Plotting function OplsMAte object
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
