% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{plot,OplsMate-method}
\alias{plot,OplsMate-method}
\title{Plotting function OplsMAte object}
\usage{
\S4method{plot}{OplsMate}(x, y)
}
\arguments{
\item{x}{OplsMAte object}

\item{y}{string, describing plot type: scores, scores_cv}
}
\value{
graphics output
}
\description{
Plotting function OplsMAte object
}
\section{}{
NA
}

\examples{
data(iris)
smod=opls(X=iris[,seq_len(4)], Y=iris$Species=='setosa')
plot(smod, 'scores_cv')
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
