% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{plot,PCA_metabom8-method}
\alias{plot,PCA_metabom8-method}
\title{Plotting function PCA}
\usage{
\S4method{plot}{PCA_metabom8}(x, an, pc, title = "", qc = NA, legend = "in", cv.scores = T, ...)
}
\arguments{
\item{x}{PCA_metabom8 object}
}
\value{
ggplot2 output
}
\description{
Plotting function PCA
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
