% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocsy.R
\name{plot_stocsy}
\alias{plot_stocsy}
\title{Plot Statistical Total Correlation Spectroscopy (STOCSY)}
\usage{
plot_stocsy(stoc_mod, shift = c(0, 10), title = NULL)
}
\arguments{
\item{title}{char, plot title}

\item{X}{Numeric matrix or dataframe where each row represents a NMR spectrum and each column a chemical shift variable.}

\item{ppm}{num array of chemical shift variables, matched to columns in X}

\item{driver}{num, chemical shift indicating STOCSY driver signal}

\item{plotting}{logical, indicing if results should be plotted}
}
\description{
Plot Statistical Total Correlation Spectroscopy (STOCSY)
}
\section{}{
NA
}

\seealso{
Other metID functions: 
\code{\link{stocsy}()}
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
\concept{metID functions}
