% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllViz.R
\name{plotscores}
\alias{plotscores}
\title{Plotting PCA or OPLS model scores}
\usage{
plotscores(
  model,
  pc = c(1, 2),
  an,
  title = "",
  qc = NA,
  legend = "in",
  cv.scores = T,
  ...
)
}
\arguments{
\item{model}{PCA, PLS or OPLS model of the package \emph{MetaboMate}.}

\item{pc}{Specifies which model components should be plotted on abscissa and ordinate (see Details).}

\item{an}{Colour and label specification (see Details).}

\item{title}{Plot title.}

\item{qc}{Row indices of QC samples. Can be left NA, however, if specified QC samples will be highlighted in the plot.}

\item{legend}{Position of the plot legend, set to NA if legend should be outside of the plotting area.}

\item{cv.scores}{Logical value (TRUE or FALSE) indicating if cross-validated scores should be plotted for the predictive component(s) (only for PLS and O-PLS).}

\item{...}{Additional values passed on to \code{\link[ggplot2]{scale_colour_gradientn}}.}
}
\value{
This function returns a \emph{ggplot2} plot object.
}
\description{
Function for plotting PCA, PLS or OPLS model scores.
}
\details{
Scores colouring is specified with the argument \code{an}, which is a list of three elements. The first list element specifies the colour (class factor required for categorical variables), the second list element specifies a point labeling (class character or factor) and the third list element specifies point shape. The Hotelling's \out{T<sup>2</sup>} ellipse is automatically included and calculated for the dimensions selected by the \code{pc} argument.
}
\section{}{
NA
}

\references{
Trygg J. and Wold, S. (2002) Orthogonal projections to latent structures (O-PLS). \emph{Journal of Chemometrics}, 16.3, 119-128.

Hotelling, H. (1931) The generalization of Student’s ratio. \emph{Ann. Math. Stat.}, 2, 360-378.
}
\author{
Torben Kimhofer \email{tkimhofer@gmail.com}

Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
