% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_stocsy.R
\name{plotStocsy}
\alias{plotStocsy}
\title{Plot STOCSY result}
\usage{
plotStocsy(stoc_mod, shift = c(0, 10), title = NULL)
}
\arguments{
\item{stoc_mod}{An object of class \code{m8_stocsy1d} returned by \code{stocsy()}.}

\item{shift}{Numeric vector of length 2 specifying the chemical shift range (ppm).}

\item{title}{Optional character plot title.}
}
\value{
A \code{ggplot2} object.
}
\description{
Generates a STOCSY plot (covariance trace coloured by absolute correlation).
}
\examples{
# st <- stocsy(X, ppm, driver = 5.233, plotting = FALSE)
# plotStocsy(st, shift = c(5.15, 5.30), title = "Glucose")

data(covid)
cs = 5.233 # anomeric H of gluc
s1 <- stocsy(covid$X, driver=cs, plotting = FALSE)
plotStocsy(s1)

}
