% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-s4_generics.R, R/03-s4_methods.R
\name{prep}
\alias{prep}
\alias{prep,m8_preprocess-method}
\title{Applies a preprocessing strategy to a numeric matrix.}
\usage{
prep(object, X)

\S4method{prep}{m8_preprocess}(object, X)
}
\arguments{
\item{object}{A preprocessing strategy object.}

\item{X}{Numeric matrix.}
}
\value{
A list containing the processed matrix and parameters.
}
\description{
Applies a preprocessing strategy to a numeric matrix.
}
\examples{
uv <- UVScaling(center=TRUE)
X <- matrix(c(10,10, 0,0, 0, 10, 0, 1000), ncol=4)
X_scaled <- prep(uv, X)
str(X_scaled)
}
