% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scalings.R
\name{prep_X}
\alias{prep_X}
\title{Applies a preprocessing strategy to a numeric matrix.}
\usage{
prep_X(preproc_strategy, X)
}
\arguments{
\item{preproc_strategy}{list with elements 'center' (logical) and 'scale' (character).}

\item{X}{numeric matrix - processing is done column-wise.}
}
\value{
A list containing the processed matrix and parameters.
}
\description{
Applies a preprocessing strategy to a numeric matrix.
}
\examples{
X <- matrix(c(10,10, 0,0, 0, 10), ncol=3)
autoscale <- uv_scaling(center=TRUE)
X_scaled <- prep_X(autoscale, X)
str(X_scaled)
}
