% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-s4_generics.R, R/03-s4_methods.R
\name{scores}
\alias{scores}
\alias{scores,m8_model-method}
\title{PLS/OPLS model scores}
\usage{
scores(object, ...)

\S4method{scores}{m8_model}(object, orth = FALSE, cv = FALSE)
}
\arguments{
\item{object}{An object of class \code{m8_model}.}

\item{...}{Additional arguments (currently ignored).}

\item{orth}{Logical indicating whether orthogonal scores should be returned
(only applicable for OPLS models).}

\item{cv}{Logical indicating whether cross-validated scores should be returned}
}
\value{
Numeric vector or matrix containing scores.
}
\description{
PLS/OPLS model scores
}
\examples{
data(covid)
cv <- balanced_mc(k=5, split=2/3)
scaling <- uv_scaling(center=TRUE)
model <-opls(X=covid$X, Y=covid$an$type, scaling, cv)
show(model)
scores(model, orth=FALSE)
scores(model, orth=TRUE)
scores(model, cv=TRUE)
}
