% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{scRange}
\alias{scRange}
\title{Min-max scaling}
\usage{
scRange(x, ra)
}
\arguments{
\item{x}{num array, desired value range of the scaled variable}
}
\value{
num array, same length as x
}
\description{
Min-max scaling
}
\details{
Output values are scaled to range between 0 and 1
}
\section{}{
NA
}

\examples{
x=rnorm(20, 0, 1)
plot(x, type='l'); abline(h=range(x), lty=2)
points(scRange(x, ra=c(5,15)), type='l', col='red'); abline(h=c(0,1), col='red', lty=2)
}
\seealso{
Other NMR ++: 
\code{\link{.hotellingsT2}()},
\code{\link{cvanova}()},
\code{\link{es_cdelta}()},
\code{\link{minmax}()},
\code{\link{opls_perm}()},
\code{\link{pca}()},
\code{\link{plotscores}()},
\code{\link{pqn}()},
\code{\link{predict_opls}()}
}
\author{
\email{torben.kimhofer@gmail.com}
}
\concept{NMR ++}
