% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllViz.R
\name{spec}
\alias{spec}
\title{Simple plotting of a single NMR spectrum}
\usage{
spec(
  x,
  ppm,
  shift = c(0, 11),
  add = F,
  interactive = T,
  name = "A",
  mode = "lines",
  ...
)
}
\arguments{
\item{x}{NMR spectrum.}

\item{ppm}{ppm vector.}

\item{shift}{chemical shift region to be plotted.}

\item{add}{Logical indicating if spectrum should be added to a current plot generated with \code{spec()} or \code{matspec()}.}

\item{interactive}{logical, interactive version (plotly package)}

\item{name}{string, name of trace (only used in interactive mode)}

\item{mode}{string, plot mode for interactive version: 'lines', 'lines+markers' or 'markers' (see Details)}

\item{...}{Additional parameters to be passed on to the graphics generic plot function.}
}
\description{
Simple plotting of a single NMR spectrum
}
\details{
Low-level plotting function for a single NMR spectrum (base graphics). Interactive visualisation in mode of markers or line+markers will take longer (mode='lines' is recommended).
}
\section{}{
NA
}

\seealso{
\code{\link{matspec}} \code{\link{plot}}

Other visNMR: 
\code{\link{matspec}()}
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
\concept{visNMR}
