% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllViz.R
\name{specOverlay}
\alias{specOverlay}
\title{Higher level plotting function to overlay NMR spectra (ggplot2 based)}
\usage{
specOverlay(
  X,
  ppm = NULL,
  shift = c(-0.01, 0.01),
  an = list("facet", "col", "ltype"),
  alp = 0.7,
  size = 0.5,
  title = "",
  ...
)
}
\arguments{
\item{X}{Input NMR data matrix with row representing spectra.}

\item{ppm}{ppm vector with its length equals to \code{nrow(X)}.}

\item{shift}{Chemical shift area to be plotted. This should be kept as small as possible (see Details).}

\item{an}{List with one to three elements specifying facetting, colour and linetype (see Details).}

\item{alp}{Alpha value for lines (number between 0 and 1 whereas 0 is fully transparent and 1 is fully opaque).}

\item{size}{Line width (0.5 is a good start).}

\item{title}{Plot title.}

\item{...}{Additional paramters passed on to ggplot's facet function.}
}
\value{
ggplot2 plot object
}
\description{
Plotting overlayed NMR specra. This function is based on ggplot2, a high-level plotting R package. For large ppm ranges the computation time is relatively long, so the chemical \code{shift} range should be as small as possible. For list argument \code{an}, the first element describes the colour and must be defined (even if it is only a single value). If colour and line width are specified, then at least one list elements of \code{an} must have the same length as \code{X}.
}
\section{}{
NA
}

\examples{
data(covid)
panel=sample(c('A', 'B'), nrow(X), replace = TRUE)
specOverlay(X, shift=c(5.15, 4.6), an=list(panel, Date=meta$a_DATE, AUNM=meta$a_AUNM))
specOverlay(X, shift=c(5.15, 5.3), an=list(panel, RunOrder=as.numeric(meta$a_DATE), AUNM=meta$a_AUNM))
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
