% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{stocsy1d_metabom8-class}
\alias{stocsy1d_metabom8-class}
\title{An S4 class to represent STOCSY model constructed with metabom8}
\description{
An S4 class to represent STOCSY model constructed with metabom8
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
