% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-s4_generics.R, R/03-s4_methods.R
\name{train_idc}
\alias{train_idc}
\alias{train_idc,ResamplingInstance-method}
\title{Extract training indices from a resampling instance}
\usage{
train_idc(object)

\S4method{train_idc}{ResamplingInstance}(object)
}
\arguments{
\item{object}{A \code{ResamplingInstance}.}
}
\value{
A list where each element contains the row indices
corresponding to the training set of a resampling split.
}
\description{
Extract training indices from a resampling instance
}
\details{
Returned is a list of training set indices stored in a
\code{ResamplingInstance} object.
}
\examples{
n <- 100
thr <- 1.5
Y <- c(rnorm(80, thr - 3, 0.3), rnorm(20, thr + 3, 0.3))  # unbalanced outcome
cv_k  <- kfold(k = 10)
k_inst <- instantiate(cv_k,  matrix(Y, ncol = 1))
idc <- train_idc(k_inst)
head(idc)
}
