% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scalings.R
\name{unscaled}
\alias{unscaled}
\title{No Scaling
This function defines a preprocessing strategy that is applied via
\code{\link{prep_X}}.}
\usage{
unscaled(center = FALSE)
}
\arguments{
\item{center}{Logical. If TRUE, variables are mean-centered before scaling.}
}
\value{
A \code{list} with elements:
\describe{
\item{X}{Numeric matrix containing the scaled data.}
\item{prep}{List describing the preprocessing, including centering and scaling
parameters (\code{center}, \code{scale}, \code{X_mean}, \code{X_sd}).}
}#' @details Leaves variables unscaled. Optional centering.
}
\description{
No Scaling
This function defines a preprocessing strategy that is applied via
\code{\link{prep_X}}.
}
\examples{
just_centering <- unscaled(center=TRUE)
X <- matrix(c(10,10, 0,0, 0, 10), ncol=3)
X_centered <- prep_X(just_centering, X)
str(X_centered)
X_centered$X
}
\seealso{
Other scaling_strategies: 
\code{\link{pareto_scaling}()},
\code{\link{uv_scaling}()}
}
\concept{scaling_strategies}
