% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scalings.R
\name{uv_scaling}
\alias{uv_scaling}
\title{Unit Variance Scaling
This function defines a preprocessing strategy that is applied via
\code{\link{prep_X}}.}
\usage{
uv_scaling(center = TRUE)
}
\arguments{
\item{center}{Logical. If TRUE, variables are mean-centered before scaling.}
}
\value{
A \code{list} with elements:
\describe{
\item{X}{Numeric matrix containing the scaled data.}
\item{prep}{List describing the preprocessing, including centering and scaling
parameters (\code{center}, \code{scale}, \code{X_mean}, \code{X_sd}).}
}
}
\description{
Unit Variance Scaling
This function defines a preprocessing strategy that is applied via
\code{\link{prep_X}}.
}
\details{
Centers variables and scales each feature to unit variance.
UV scaling divides each variable by its standard deviation.
This is the default scaling in many multivariate methods such as PCA and PLS.
}
\examples{
autoscale <- uv_scaling(center=TRUE)
X <- matrix(c(10,10, 0,0, 0, 10), ncol=3)
X_scaled <- prep_X(autoscale, X)
str(X_scaled)
X_scaled$X
}
\seealso{
Other scaling_strategies: 
\code{\link{pareto_scaling}()},
\code{\link{unscaled}()}
}
\concept{scaling_strategies}
