% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-s4_classes.R
\name{UVScaling}
\alias{UVScaling}
\title{Unit Variance Scaling
Centers variables and scales each feature to unit variance.}
\usage{
UVScaling(center = TRUE)
}
\arguments{
\item{center}{Logical. If TRUE, variables are mean-centered before scaling.}
}
\value{
An object of class \code{m8_preprocess}.
}
\description{
Unit Variance Scaling
Centers variables and scales each feature to unit variance.
}
\details{
UV scaling divides each variable by its standard deviation.
This is the default scaling in many multivariate methods such as PCA and PLS.
}
\examples{
autoscale <- UVScaling(center=TRUE)
X <- matrix(c(10,10, 0,0, 0, 10), ncol=3)
X_scaled <- prep(autoscale, X)
str(X_scaled)
X_scaled$X
}
\seealso{
Other scaling objects & strategies: 
\code{\link{NoScaling}()},
\code{\link{ParetoScaling}()}
}
\concept{scaling objects & strategies}
