% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-s4_generics.R, R/03-s4_methods.R
\name{vip}
\alias{vip}
\alias{vip,m8_model-method}
\title{Variable Importance in Projection (VIP)}
\usage{
vip(object)

\S4method{vip}{m8_model}(object)
}
\arguments{
\item{object}{An object of class \code{m8_model}.}
}
\value{
Numeric vector or matrix containing the variable importance in
projection (VIP) scores.
}
\description{
Variable Importance in Projection (VIP)
}
\examples{
data(covid)
cv <- balanced_mc(k=5, split=2/3)
scaling <- uv_scaling(center=TRUE)
model <-opls(X=covid$X, Y=covid$an$type, scaling, cv)
show(model)
vip_scores <- vip(model)
dim(vip_scores)
}
