% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignSegment.R
\name{alignSegment}
\alias{alignSegment}
\title{Align NMR segment to refrerence based on maximum cross-correlation}
\usage{
alignSegment(seg, idx_ref = 1, clim = 0.7, med = TRUE, norm = FALSE)
}
\arguments{
\item{seg}{matrix or data frame, NMR submatrix containing segments (spectra in rows)}

\item{idx_ref}{int, rown index of reference spectrum}

\item{clim}{num, upper cross-correlation threshold (see Details)}

\item{med}{logical, use median spectrum as reference?}

\item{norm}{logical, minmax normalisation for each spectrum prior to alignment? Useful for viz.}
}
\value{
Matrix of aligned spectral segments
}
\description{
Align NMR segment to refrerence based on maximum cross-correlation
}
\details{
Spectral alignment is obtained by x-positional shifting a spectral segment. The extend of the shift is determined by maximising the cross-correlation with a refrence segment. The refernce can either be a spectrum in \emph{seg} or the segment median spectrum (parameter \emph{idx_ref} is ignored then). Non-matching spectral segments, defined by \emph{clim} (cross-corrleation upper limit, similarity score) are not aligned.
}
\section{}{
NA
}

\seealso{
Other NMR: 
\code{\link{bline}()},
\code{\link{excise1d}()},
\code{\link{get.idx}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{normqref}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{spec}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{stocsy}()},
\code{\link{storm}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR}
