% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\title{Spectral data binning}
\usage{
binning(X, ppm, width = NULL, npoints = NULL)
}
\arguments{
\item{X}{num matrix, NMR data with spectra in rows}

\item{ppm}{num array, chemical shift positions, length matches to columns in X}

\item{width}{num, bin size in ppm or NULL in case \code{npoints} is specified}

\item{npoints}{num, desired number of bins per spectrum or NULL in case \code{width} is specified}
}
\value{
Numeric matrix with spectra in rows and chemical shift variables in columns.
}
\description{
Spectral data binning
}
\details{
Equidistant binning of spectra. Specify either \code{width} or \code{npoints} argument - if both are provided, \code{npoints} is used. Input argument \code{ppm} can be omitted if chemical shift information is encoded in the column names of the NMR matrix \code{X}.
}
\examples{
# Simulate 2 spectra with 100 ppm points
set.seed(1)
X <- matrix(rnorm(2 * 100), nrow = 2)
ppm <- round(seq(10, 0.5, length.out = 100), 3)
colnames(X) <- ppm

# Bin using width of 0.5 ppm
Xb <- binning(X, ppm, width = 0.5)

# Show dimensions of binned data
dim(Xb)
}
\seealso{
Other NMR: 
\code{\link{alignSegment}()},
\code{\link{cvanova}()},
\code{\link{get_idx}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{normErectic}()},
\code{\link{read1d_raw}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{storm}()}
}
\concept{NMR}
