% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\title{Equidistant binning of spectra}
\usage{
binning(X, ppm, width = 0.001, npoints = NULL)
}
\arguments{
\item{X}{NMR matrix (spectra in rows)}

\item{ppm}{num ppm vector}

\item{width}{num, bin size}

\item{npoints}{num, desired number of bins per spectrum}
}
\value{
Use either width or npoints argument. Function returns is a matrix (rows = spectra, columns=bins).
}
\description{
Equidistant binning of spectra
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
