% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\title{NMR data binning
Equidistant binning of spectra based on either a specified bin witdth or the number of desired bins.}
\usage{
binning(X, ppm, width = 0.001, npoints = NULL)
}
\arguments{
\item{X}{num matrix, NMR data with spectra in rows}

\item{ppm}{num array, chemical shift positions, length matches to columns in X}

\item{width}{num, bin size in ppm or NULL in case \code{npoints} is specified}

\item{npoints}{num, desired number of bins per spectrum or NULL in case  \code{width} is specified}
}
\value{
Numeric matrix whith spectra in rows and chemical shift information in columns.
}
\description{
NMR data binning
Equidistant binning of spectra based on either a specified bin witdth or the number of desired bins.
}
\details{
Specify either \code{width} or \code{npoints} argument - if both are fiven, \code{npoints} is used. Input argument \code{ppm} can be omitted if chemical shift infomration is encoded in the column names of the NMR matrix \code{X}. The returne matrix encodes chemical shift information as column names.
}
\examples{
load(covid)
Xb=binning(X, ppm, width=0.005)
ppm_bin=as.numeric(colnames(Xb))
par(mfrow=c(2,1))
spec(X[1,], ppm, shift=c(5.15, 5.3), interactive=FALSE)
spec(Xb[1,], ppm_bin,  shift=c(5.15, 5.3), interactive=FALSE)
}
\seealso{
Other NMR: 
\code{\link{bline}()},
\code{\link{get.idx}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{spec}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{stocsy}()}
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR}
