% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bline.R
\name{bline}
\alias{bline}
\title{Baseline correction for 1D NMR spectra}
\usage{
bline(X)
}
\arguments{
\item{X}{NMR data matrix or dataframe with rows representing spectra.}
}
\value{
Baseline corrected NMR matrix.
}
\description{
Baseline correction for NMR spectra. This function firstly estimates a baseline trend based on a asymmetric least squares. Secondly the estimated baseline is subtracted from the NMR spectrum. Baseline estimation uses the the \code{asysm} function as implemented in the \code{ptw} package.
}
\seealso{
\code{\link{readBruker}} \code{\link[ptw]{asysm}}
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
