% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bline.R
\name{bline}
\alias{bline}
\alias{blcor}
\title{Baseline correction for 1D NMR spectra}
\usage{
bline(X)
}
\arguments{
\item{X}{num matrix or data.frame, NMR data with rows representing spectra}
}
\value{
Baseline corrected NMR matrix an with same dimension as input matrix.
}
\description{
Non-linear baseline correction for NMR spectra. This function estimates smooth and non-linear baseline trends for each spectrum based on a asymmetric least squares method. The estimated baseline is subtracted from each spectrum and the remainder is returned.
}
\seealso{
\code{\link[ptw]{asysm}}

Other NMR: 
\code{\link{binning}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR}
