% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bline.R
\name{bline}
\alias{bline}
\alias{bcor}
\title{Baseline Correction for 1D NMR Spectra}
\usage{
bline(X, lambda = 1e+07, iter_max = 30)

bcor(X, lambda = 1e+07, iter_max = 30)
}
\arguments{
\item{X}{Numeric matrix or data frame. NMR data with spectra in rows.}

\item{lambda}{Numeric. Smoothing parameter passed to \code{\link[ptw]{asysm}}. Larger values result in smoother baselines.}

\item{iter_max}{Integer. Maximum number of iterations for the baseline estimation algorithm.}
}
\value{
A numeric matrix of the same dimensions as the input, containing the baseline-corrected spectra.
}
\description{
Non-linear baseline correction for NMR spectra based on asymmetric least squares. This function estimates a smooth, non-linear baseline trend for each spectrum. The estimated baseline is then subtracted, returning the corrected spectrum. See \code{\link[ptw]{asysm}} for more information on the smoothing parameter \code{lambda}.
}
\details{
Missing values in \code{X} are replaced with zeros before baseline correction.
}
\examples{
data(covid_raw)
X1_bc <- bcor(X[1, ])
spec(X[1, ], ppm, shift = c(3, 4), interactive = FALSE)
spec(X1_bc[1, ], ppm, shift = c(3, 4), add = TRUE, interactive = FALSE)

}
\seealso{
\code{\link[ptw]{asysm}}, \code{\link{spec}}, \code{\link{matspec}}
}
