% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bline.R
\name{bline}
\alias{bline}
\alias{blcor}
\title{Baseline correction for 1D NMR spectra}
\usage{
bline(X)
}
\arguments{
\item{X}{num matrix or data.frame, NMR data with rows representing spectra}
}
\value{
Baseline corrected NMR matrix an with same dimension as input matrix.
}
\description{
Non-linear baseline correction for NMR spectra. This function estimates smooth and non-linear baseline trends for each spectrum based on a asymmetric least squares method. The estimated baseline is subtracted from each spectrum and the remainder is returned.
}
\examples{
data(covid_raw)
X1_bc=bline(X[1,])
spec(X[1,], ppm, shift=c(3,4))
spec(X1_bc[1,], ppm, shift=c(3,4), add=TRUE)
}
\seealso{
\code{\link[ptw]{asysm}}

Other NMR: 
\code{\link{binning}()},
\code{\link{get.idx}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{spec}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{stocsy}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR}
