% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Spectral calibration to a chemical shift reference (1D NMR)}
\usage{
calibrate(X, ppm, type = c("tsp", "glucose", "alanine"))
}
\arguments{
\item{X}{num matrix, NMR data matrix with rows representing spectra and columns representing chemical shift position}

\item{ppm}{num vector, matched to columns of X}

\item{type}{str or num. Str: Either 'tsp' or 'glucose' for urine or blood-derived spectra, respectively (see Details). Num: ppm range of max height signal that will be used to reference to zero}
}
\value{
num matrix X, calibrated NMR data matrix.
}
\description{
Spectral calibration to a chemical shift reference.
}
\details{
Spectral calibration to a selected chemical shift reference signal.
\itemize{
\item{\code{type='tsp'}{calibration to 0 ppm using the highest peak located in interval 0 +/- 0.20 ppm (Trimethylsilylpropanoic acid resonance)}}
\item{\code{type='glucose'}{calibration to 5.23 ppm using the glucose doublet located in interval 5.15 ppm to 5.30 ppm}}
\item{\code{type='alanine'}{calibration to 1.48 ppm using the alanine doublet located in interval 1.4 ppm to 1.53 ppm}}
\item{\code{type=ppm_range}{ppm_range is numeric array, calibration to mean(ppm_range) using a maximum signal located in interval ppm_range}}
}
}
\examples{
data(covid_raw)
matspec(X, ppm, shift=c(-0.1,0.1))
X_tsp=calibrate(X, ppm, type='tsp')
matspec(X, ppm, shift=c(-0.1,0.1))

# calibrate with glucose (blood plasma)
matspec(X, ppm, shift=c(5.15, 5.3))
X_tsp=calibrate(X, ppm, type='glucose')
matspec(X, ppm, shift=c(5.15, 5.3))

}
\references{
Dona, A.C., \emph{et al.} (2014) Precision high-throughput proton NMR spectroscopy of human urine, serum, and plasma for large-scale metabolic phenotyping. \emph{Analytical Chemistry}. 86.19. 9887-94.
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
