% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllViz.R
\name{dmodx}
\alias{dmodx}
\title{Distance to the Model in X-Space (DModX)}
\usage{
dmodx(mod, plot = TRUE)
}
\arguments{
\item{mod}{An OPLS model object of class \code{OPLS_metabom8}.}

\item{plot}{Logical. If \code{TRUE}, a plot of DModX values with a cutoff line based on a t-test is shown.}
}
\value{
A data frame with columns:
\describe{
\item{ID}{Sample index}
\item{DmodX}{Distance to the model in X-space}
\item{passedT.test}{Logical. TRUE if within the 95\\\% confidence interval}
}
}
\description{
Calculates the orthogonal distance of each observation to the OPLS model in X-space. The DModX is used for identifying outliers.
}
\details{
DModX is calculated as the scaled root-mean-squared residual. An approximate upper 95\\\% confidence limit is drawn using a one-sided t-test.
}
\examples{
data(covid)
X <- covid$X
an <- covid$an
model <- opls(X, Y = an$type)
dmx <- dmodx(model)

}
\references{
Bylesjo, M. et al. (2006). \emph{J. Chemometrics}, 20, 341–351.
Wold, S. (1976). \emph{Pattern Recognition}, 8, 127–139.
}
\seealso{
\code{\link{opls}}, \code{\link{plotload}}
}
