% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate1d_gluc.R
\name{.calibrate1d_gluc}
\alias{.calibrate1d_gluc}
\title{Calibrate 1D to glucose @ 5.233 (d)}
\usage{
.calibrate1d_gluc(X, ppm)
}
\arguments{
\item{X}{num matrix, NMR matrix with spectra in rows}

\item{ppm}{num vec, chemical shift matching to X}
}
\value{
Calibrated X matrix
}
\description{
Calibrate 1D to glucose @ 5.233 (d)
}
\author{
Torben Kimhofer \email{torben.kimhofer@gmail.com}
}
\keyword{internal}
