% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_doub.R
\name{.calibrate_doub}
\alias{.calibrate_doub}
\title{Calibrate spectra to a doublet signal}
\usage{
.calibrate_doub(
  X,
  ppm,
  type = c("glu", "ala"),
  j_const = c(0.006, 0.007),
  sg_length = 13
)
}
\arguments{
\item{X}{num matrix, NMR matrix with spectra in rows}

\item{ppm}{num vec, chemical shift matching to X}

\item{type}{string/num vec, either glu or ala for glucose and alanine calibration, respectively, or num array defining chem shift of doublet to be used for calibration}

\item{j_const}{num array, range of J constant values used to identify doublet signal (ppm), default is for glucose (doublet has J-cons of 3.85 Hz)}

\item{sg_length}{int, nb of data points to cacl 2nd degree estimation with savitzki-golay (higher number -> smoother)}
}
\value{
Calibrated X matrix
}
\description{
Calibrate spectra to a doublet signal
}
\author{
Torben Kimhofer \email{torben.kimhofer@gmail.com}
}
\keyword{internal}
