% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1d_raw.R
\name{.check1d_files_fid}
\alias{.check1d_files_fid}
\title{Check for intact file systems - helper function read1d}
\usage{
.check1d_files_fid(datapath, n_max = 10, filter = T)
}
\arguments{
\item{datapath}{char, File directory containing spectra}

\item{n_max}{int, Maximum number of spectra to read-in}

\item{filter}{lobic, filter for intact file systems (TRUE is recommended)}

\item{procs_exp}{num, Topspin processing experiment ID}
}
\description{
Check for intact file systems - helper function read1d
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
