% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.check_X_ppm}
\alias{.check_X_ppm}
\title{Check dimension, NA and infinity}
\usage{
.check_X_ppm(X, ppm)
}
\arguments{
\item{X}{NMR matrix with spectra represented in rows.}

\item{ppm}{ppm vector.}
}
\value{
Logical indicating if X and ppm match and absence of NA and inf in ppm
}
\description{
Check dimension, NA and infinity
}
\section{}{
NA
}

\keyword{internal}
