% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.checkDimXY}
\alias{.checkDimXY}
\title{Check X and Y shapes (PLS context)}
\usage{
.checkDimXY(X, Y)
}
\arguments{
\item{X}{X matrix (observations x variables)}

\item{Y}{Y matrix (observations x variables)}
}
\value{
NULL if all's fine and throws error otherwise
}
\description{
This function checks for matching dimensions (X - Y).
}
\section{}{
NA
}

\keyword{internal}
