% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.checkYclassNas}
\alias{.checkYclassNas}
\title{Check for missing values (PLS context)}
\usage{
.checkYclassNas(Y)
}
\arguments{
\item{Y}{Input data (uni or multivar) formatted as vector or matrix or data.frame}
}
\value{
List of i) Y matrix, ii) Y levels (empty if numeric Y), iii) Y type (R or DA)
}
\description{
This function checks for missing values in Y (NA, NAN, infinite) and establishes analysis type accroding to the class of the input data Y: regression (R) or discriminant analysis (DA). It also converts Y to a matrix.
}
\section{}{
NA
}

