% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.checkYclassNas}
\alias{.checkYclassNas}
\title{Check Y for missing values (PLS context)}
\usage{
.checkYclassNas(Y)
}
\arguments{
\item{Y}{A vector, matrix, or data.frame. Target variable.}
}
\value{
A list: cleaned Y matrix, levels (if applicable), and type ('R' or 'DA')
}
\description{
Checks input Y for NA/NaN/Inf values, determines analysis type
(Regression or Discriminant Analysis), and returns cleaned Y matrix,
associated levels, and type.
}
\keyword{internal}
