% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.cvSetsMethod}
\alias{.cvSetsMethod}
\title{Create Cross Validation (CV) row-indices for Y}
\usage{
.cvSetsMethod(
  Y,
  type = c("R", "DA"),
  method = "k-fold_stratified",
  k = 7,
  split = 2/3
)
}
\arguments{
\item{Y}{matrix, CV row indices to be generated for}

\item{type}{char Regression (R) or disciminant analysis (DA)}

\item{method}{Type of cross validation: k-fold or Monte Carlo-Cross Validation (MCCV), sampling can be performed group-blind (\code{'k-fold'}, \code{'MC'}) or group-stratified (\code{'k-fold_stratified'}, \code{'MC_balanced'}).}

\item{k}{Number of training sets to generate.}

\item{split}{Fraction of samples used to generate training set for class/group-balanced Monte Carlo (MC) CV}
}
\value{
list of Y row-indices for each CV fold
}
\description{
This function creates a list of indices that represent k fold subsets for MCCV
}
\section{}{
NA
}

