% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1d.R
\name{.detect1d_procs}
\alias{.detect1d_procs}
\title{Check for intact file systems - helper function read1d!}
\usage{
.detect1d_procs(datapath, n_max = 10, filter = TRUE, recursive, verbose)
}
\arguments{
\item{datapath}{char, File directory containing spectra}

\item{n_max}{int, Maximum number of spectra to read-in}

\item{filter}{lobic, filter for intact file systems (TRUE is recommended)}

\item{procs_exp}{num, Topspin processing experiment ID}
}
\description{
Check for intact file systems - helper function read1d!
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
\keyword{internal}
