% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1d.R
\name{.detect1d_procs}
\alias{.detect1d_procs}
\title{Check for intact file systems - helper function read1d!}
\usage{
.detect1d_procs(datapath, n_max = 10, filter = TRUE, recursive, verbose)
}
\arguments{
\item{datapath}{char, File directory containing spectra}

\item{n_max}{int, Maximum number of spectra to read-in}

\item{filter}{lobic, filter for intact file systems (TRUE is recommended)}

\item{procs_exp}{num, Topspin processing experiment ID}
}
\value{
list of file descriptors: experiment folder path, folder id, procs file, acqus file, 1r file.
list(path=p_intact, exp_no=exp_no, f_procs=f_procs, f_acqus=f_acqus, f_1r=f_1r)
}
\description{
Check for intact file systems - helper function read1d!
}
\section{}{
NA
}

\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\keyword{internal}
