% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.dimX}
\alias{.dimX}
\title{Ensure Input is a Row Matrix}
\usage{
.dimX(X)
}
\arguments{
\item{X}{Numeric vector, matrix, or array. Typically NMR data where rows represent spectra.}
}
\value{
A numeric matrix with one row if input was a vector; otherwise, returns \code{X} unchanged.
}
\description{
Ensures that the input object \code{X} is in row-matrix form. If \code{X} is a vector (i.e., no \code{ncol} attribute),
it is converted to a matrix with 1 row and \code{length(X)} columns.
This is helpful for standardizing inputs before applying matrix operations.
}
\keyword{internal}
