% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.extMeanCvFeat}
\alias{.extMeanCvFeat}
\title{Extract data structures from output of function .oplsComponentCv}
\usage{
.extMeanCvFeat(cv_obj, feat = "t_xp", cv_type, model_type)
}
\arguments{
\item{cv_obj}{named list, output of .oplsComponentCv (for resp OPLS component)}

\item{feat}{char, feature name in output of tt (names(tt))}

\item{cv_type}{char indicating cross validation type (k-fold, k-fold_stratified, MC, MC_balanced)}

\item{model_type}{char indicating single or multi-column Y (grep mY)}
}
\value{
num array or matrix of resp component feature. For MCCV and single-column (multi-column) Y this is a 2D (3D) matrix where D1: mean, sd, or length, D2: nrow(Y) and D3: ncol(Y). For k-fold CV single-column (multi-column) Y output is 1D (2D) with D1: nrow(Y) and D2: ncol(Y)
}
\description{
Extract data structures from output of function .oplsComponentCv
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
