% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawGenerics.R
\name{.extract_acq_pars1d}
\alias{.extract_acq_pars1d}
\title{Extract Bruker NMR Acquisition Parameters}
\usage{
.extract_acq_pars1d(f_list)
}
\arguments{
\item{f_list}{List. Contains paths to NMR experiment folders (e.g., as returned by \code{.check1d_files_fid()}).}
}
\value{
A data frame containing extracted acquisition parameters for each experiment.
}
\description{
Parses \code{acqus} files to extract acquisition parameters for each experiment.
}
\seealso{
\code{\link{.filterExp_files}}, \code{\link{.check1d_files_fid}}
}
\keyword{internal}
