% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1d.R
\name{.extract_pars1d}
\alias{.extract_pars1d}
\title{Read Bruker NMR paramter files - helper function read1d}
\usage{
.extract_pars1d(f_list)
}
\arguments{
\item{f_list}{list, intact files system for NMR experiments. See fct checkFiles1d}
}
\description{
Read Bruker NMR paramter files - helper function read1d
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
\keyword{internal}
