% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawGenerics.R
\name{.fidApodisationFct}
\alias{.fidApodisationFct}
\title{modGauss'={},
expDampJmod'={},
sineShift'={if( all(c('offs', 'end', 'exp') \%in\% names(pars))){afun<-.sineMod(n, pars$offs, pars$end, pars$exp)} else{stop('Check aposation fct arguments: sineShift function requires offs, end, exp parameters')}},
triangle'={},}
\usage{
.fidApodisationFct(n, pars)
}
\description{
modGauss'={},
expDampJmod'={},
sineShift'={if( all(c('offs', 'end', 'exp') \%in\% names(pars))){afun<-.sineMod(n, pars$offs, pars$end, pars$exp)} else{stop('Check aposation fct arguments: sineShift function requires offs, end, exp parameters')}},
triangle'={},
}
