% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawGenerics.R
\name{.filterExp_files}
\alias{.filterExp_files}
\title{Filter Bruker NMR Experiments}
\usage{
.filterExp_files(pars, exp_type, f_list, n_max)
}
\arguments{
\item{pars}{Data frame. Parsed parameter data from acquisition files.}

\item{exp_type}{Named list. Filtering conditions for acquisition parameters.}

\item{f_list}{List. File paths of spectra and metadata (e.g., f_fid, f_1r).}

\item{n_max}{Integer. Maximum number of experiments to retain (mainly for debugging).}
}
\value{
A list with filtered and sorted \code{f_list} and \code{pars}.
}
\description{
Filters experiments based on acquisition parameters and limits the number returned.
}
\seealso{
\code{\link{.extract_acq_pars1d}}, \code{\link{.check1d_files_fid}}
}
\keyword{internal}
