% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotellingsT2.R
\name{.hotellingsT2}
\alias{.hotellingsT2}
\title{Hotelling's T2 ellipse in 2D}
\usage{
.hotellingsT2(x, y, alpha = 0.95)
}
\arguments{
\item{x}{num vector descrbing dimension 1}

\item{y, }{num vector descrbing dimension 2}

\item{alpha}{num, probability level}
}
\value{
data.frame containing H.T2 ellipse cooredinates
}
\description{
This function is used to calculate hotellings T2 ellipse
}
\references{
Geladi, P and Kowalski, B.R. (1986), Partial least squares and regression: a tutorial. \emph{Analytica Chimica Acta}, 185, 1-17.
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
