% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.kFold}
\alias{.kFold}
\title{k-fold subsets for cross validation (CV)}
\usage{
.kFold(k, Y)
}
\arguments{
\item{k}{int k parameter}

\item{Y}{matrix (observations times variables)}
}
\value{
list of Y row-indices for each fold
}
\description{
This function ccreates a list of indices that represent k fold subsets for CV
}
\section{}{
NA
}

