% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.load_backscaled_nmr}
\alias{.load_backscaled_nmr}
\title{Backscaling for NMR data}
\usage{
.load_backscaled_nmr(mod, pc, idx, ppm)
}
\value{
data frame with backscaled loadings
}
\description{
Backscaling for NMR data
}
\section{}{
NA
}

\keyword{internal}
