% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.load_stat_reconstr_nmr}
\alias{.load_stat_reconstr_nmr}
\title{cor / cov model scores and NMR data}
\usage{
.load_stat_reconstr_nmr(mod, pc, X, idx, ppm)
}
\value{
data frame with column cor, cov, ppm
}
\description{
cor / cov model scores and NMR data
}
\section{}{
NA
}

\keyword{internal}
