% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.mc}
\alias{.mc}
\title{k-fold Monte Carlo Cross Validation (MCCV) subsets}
\usage{
.mc(k, Y, split)
}
\arguments{
\item{k}{int k parameter}

\item{Y}{matrix (observations times variables) with ncol(Y)=1, always (column reduction for multicolumn Y  in parent function: .cvSetsMethod)!}

\item{split}{Ratio of the number of samples in training and test set}
}
\value{
list of Y row-indices for each fold
}
\description{
This function creates a list of indices that represent k fold subsets for MCCV
}
\section{}{
NA
}

\keyword{internal}
