% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.oplsComponentCv}
\alias{.oplsComponentCv}
\title{Performs OPLS modelling for each CV set and collates output}
\usage{
.oplsComponentCv(X, Y, cv.set, nc, mod.cv)
}
\arguments{
\item{X}{num matrix X: preproc NMR data}

\item{Y}{num matrix Y: outcome, dummy matrix in case of categorical outcome}

\item{cv.set}{list of k elements containing vector of X and Y row-indices representing training set for cv round k}

\item{nc}{int, max number of orthogonal components to fit}

\item{mod.cv}{cv parameters}
}
\value{
Named list of collated OPLS data for respective component
}
\description{
Performs OPLS modelling for each CV set and collates output
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
\keyword{internal}
