% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{.orthModelCompSummary}
\alias{.orthModelCompSummary}
\title{Summarise component indices}
\usage{
.orthModelCompSummary(type, r2x_comp, r2_comp, q2_comp, aucs_tr, aucs_te, cv)
}
\arguments{
\item{type}{char, OPLS type: DA (discriminant analysis) or R (regression), with prefix '-mY' if mutli-column Y}

\item{r2x_comp}{num array, r2x value for each component}

\item{r2_comp}{num array, r2y value for each component}

\item{q2_comp}{num array, q2 value for each component}

\item{aucs_tr}{num array, cv training set aucs value for each component (defined only for DA)}

\item{aucs_te}{num array, cv test set aucs value for each component (defined only for DA)}

\item{cv}{list, see cv argument in function opls}
}
\value{
list: 1. data.frame, summary table 2. ggplot2, viz of summary tbl
}
\description{
Summarise component indices
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
\keyword{internal}
