% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eruption.R
\name{eruption}
\alias{eruption}
\title{Eruption plot}
\usage{
eruption(mod, pc = 1, p_adj = "BH")
}
\arguments{
\item{mod}{OPLS model generated with \code{metabom8}.}

\item{pc}{num or char, indicating OPLS component to visualise (1 for predictive component, prefix 'o' for orthogonal components, e.g., pc="o1" for first orthogonal component)}

\item{p_adj}{str, p value adjustment method, see \code{?p.adjust}}
}
\value{
gglot2 obj
}
\description{
Visualise variable importances in OPLS context, p value is generated with Kruskal Walis rank sum test, effect size measure is Cliff's delta.
}
\section{}{
NA
}

\seealso{
\code{\link{opls}}
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
