% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_cdelta.R
\name{es_cdelta}
\alias{es_cdelta}
\title{Cliff's delta (effect size)}
\usage{
es_cdelta(ref, comp)
}
\arguments{
\item{ref}{num, reference group}

\item{comp}{num, comparator group}
}
\value{
This function returns Cliff's delta effect size ranging from -1 to 1
}
\description{
Non-parametric effect size estimate used to quantify the degree to which two variable distributions overlap. Output is scaled to range between -1 to 1, with the arithmetic sign indicating if the comparator group lies to the right (higher) or left (lower) to a the reference.
}
\examples{
# define two distrubutions
a=rnorm(100, mean=0, sd=1)
b=rnorm(100, mean=2, sd=1)

# Cliff's delta
 es_cdelta(ref=a, comp=b)
 es_cdelta(ref=b, comp=a)
}
\references{
Cliff, N (1993), Dominance statistics: Ordinal analyses to answer ordinal quetsions \emph{Psychological Bulletin}, 114.(3), 494-509.\url{https://doi.org/10.1037\%2F0033-2909.114.3.494}
}
\seealso{
Other NMR ++: 
\code{\link{.hotellingsT2}()},
\code{\link{cvanova}()},
\code{\link{minmax}()},
\code{\link{opls_perm}()},
\code{\link{pca}()},
\code{\link{plotscores}()},
\code{\link{pqn}()},
\code{\link{predict_opls}()},
\code{\link{scRange}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}

\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR ++}
