% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_cdelta.R
\name{es_cdelta}
\alias{es_cdelta}
\title{Cliff's delta (effect size)}
\usage{
es_cdelta(ref, comp)
}
\arguments{
\item{ref}{num, reference group}

\item{comp}{num, comparator group}
}
\value{
This function returns Cliff's delta effect size ranging from -1 to 1
}
\description{
Non-parametric effect size estimate used to quantify the degree to which one population's variable distribution lies to the right or left of a reference group distribution.
}
\examples{
# define two distrubutions
ref=rnorm(100, mean=0, sd=1)
comp=rnorm(100, mean=2, sd=1)

# Cliff's delta
 es_cdelta(comp, ref)
 es_cdelta(ref, comp)
}
\references{
Cliff, N (1993), Dominance statistics: Ordinal analyses to answer ordinal quetsions \emph{Psychological Bulletin}, 114.(3), 494-509.\url{https://doi.org/10.1037\%2F0033-2909.114.3.494}
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}

Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
