% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excise1d.R
\name{excise1d}
\alias{excise1d}
\title{Excise standard chemical shift regions}
\usage{
excise1d(X, ppm)
}
\arguments{
\item{X}{num matrix:  NMR data, spectra in rows}

\item{ppm}{ppm, num array - chemical shift positions, length matches to columns in X}
}
\value{
The function exports the following two objects into the currently active R environment (no variable assignments needed):
\itemize{
\item Xc, num matrix: column reduced
\item ppc, num array - chemical shift positions, length is equal to ncol(X_c)

}
}
\description{
Excise standard chemical shift regions
}
\details{
Function removes the following chemical shift reagions from 1H NMR spectra acquired for small molecule research:
\itemize{
\item min(ppm) to 0.25 ppm  (cap upfield)
\item 4.5 - 5.2 ppm (residual water)
\item 5.5 - 6.0 ppm (urea)
\item 9.7 to max(ppm)  (cap downfield)
}
}
\section{}{
NA
}

\seealso{
Other NMR: 
\code{\link{alignSegment}()},
\code{\link{bline}()},
\code{\link{get.idx}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{normErectic}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{spec}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{stocsy}()},
\code{\link{storm}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR}
