% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{get.idx}
\alias{get.idx}
\alias{get_idx}
\title{Select indices of the chemical shift array taht corresponding to a chemical shift region}
\usage{
get.idx(range = c(1, 5), ppm)

get_idx(range = c(1, 5), ppm)
}
\arguments{
\item{range}{num array(2), range of chemical shift}

\item{ppm}{num array, chemical shift vector}
}
\value{
array int, indices of ppm for shift range
}
\description{
Select indices of the chemical shift array taht corresponding to a chemical shift region
}
\section{}{
NA
}

\examples{
load(covid)
idx_tsp=get.idx(c(-0.1, 0.1), ppm)
ppm[range(idx_tsp)]
plot(ppm[idx_tsp], X[1,idx_tsp], type='l')
}
\seealso{
Other NMR: 
\code{\link{alignSegment}()},
\code{\link{bline}()},
\code{\link{excise1d}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{normErectic}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{spec}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{stocsy}()},
\code{\link{storm}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR}
