% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{get.idx}
\alias{get.idx}
\title{Return indices of chemical shift array mapping to a specific chemical shift range}
\usage{
get.idx(range = c(1, 5), ppm)
}
\arguments{
\item{range}{num array(2), range of chemical shift}

\item{ppm}{num array, chemical shift vector}
}
\value{
array int, indices of ppm for shift range
}
\description{
Return indices of chemical shift array mapping to a specific chemical shift range
}
\section{}{
NA
}

\examples{
load(covid)
idx=get.idx(c(-0.1, 0.1), ppm)
ppm[range(idx)]
plot(ppm[idx], X[1,idx], type='l')
lw(X, ppm) * meta$a_SFO1 # in Hz
}
\seealso{
Other NMR: 
\code{\link{alignSegment}()},
\code{\link{binning}()},
\code{\link{bline}()},
\code{\link{excise1d}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{spec}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{stocsy}()}
}
\author{
\email{tkimhofer@gmail.com}
}
\concept{NMR}
