% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-s4_classes.R
\name{kfold}
\alias{kfold}
\title{K-fold cross-validation strategy}
\usage{
kfold(k)
}
\arguments{
\item{k}{Integer number of folds.}
}
\value{
A \code{ResamplingStrategy} object.
}
\description{
K-fold cross-validation strategy
}
\examples{
n <- 100
thr <- 1.5
Y <- c(rnorm(80, thr - 3, 0.3), rnorm(20, thr + 3, 0.3))  # unbalanced outcome
mean(Y > thr)
cv_k  <- kfold(k = 10)
k_inst <- instantiate(cv_k,  matrix(Y, ncol = 1))
sapply(train_idc(k_inst), function(i) length(i))
}
\seealso{
Other resampling strategies: 
\code{\link{balanced_boot}()},
\code{\link{balanced_mc}()},
\code{\link{instantiate}()},
\code{\link{mc}()},
\code{\link{stratified_kfold}()}
}
\concept{resampling strategies}
