% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{lw}
\alias{lw}
\title{Calculating full width at half max}
\usage{
lw(X, ppm, shift = c(-0.01, 0.01))
}
\arguments{
\item{X}{NMR matrix with rows representing spectra.}

\item{ppm}{ppm vector with its length equals to \code{nrow(X)}.}

\item{shift}{Signal shift to calculate line width}
}
\value{
Array of line widths in ppm. To convert from ppm to Hertz (Hz), multiply values with the spectrometer frequency (column \code{a_SF01} in \code{meta} data frame).
}
\description{
Calculating full width at half maximum (FWHM, aka line width). This function simply returns the ppm difference where peak line crosses half of the peak height. It requires one signal across all spectra within ppm ranges specified in \code{shift}.
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
