% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllViz.R
\name{matspec}
\alias{matspec}
\title{Plot overlayed NMR spectra}
\usage{
matspec(X, ppm, shift = c(0, 9.5), interactive = TRUE, ...)
}
\arguments{
\item{X}{num matrix, NMR matrix with spectra represented in rows.}

\item{ppm}{num array, ppm vector.}

\item{shift}{num array, hemical shift region to be plotted (in ppm).}

\item{interactive}{logical, interactive version (plotly package)}

\item{...}{additional parameters to be passed on to the graphics generic plot function.}
}
\description{
Plot overlayed NMR spectra
}
\details{
Non-interactive mode: Base graphics for NMR spectra. Interactive mode: Plotly graphics.
}
\section{}{
NA
}

\examples{
data(covid)
matspec(X[1:2,], ppm) # interactive
matspec(X[1:2,], ppm, interactive=FALSE) # static
}
\seealso{
\code{\link{spec}} \code{\link{plot}}

Other NMR: 
\code{\link{alignSegment}()},
\code{\link{bline}()},
\code{\link{excise1d}()},
\code{\link{get.idx}()},
\code{\link{lw}()},
\code{\link{noise.est}()},
\code{\link{normErectic}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{spec}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{stocsy}()},
\code{\link{storm}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR}
