% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllViz.R
\name{matspec}
\alias{matspec}
\title{Simple plotting of multiple NMR spectra overlayed}
\usage{
matspec(X, ppm, shift = c(0, 9.5), interactive = TRUE, ...)
}
\arguments{
\item{X}{NMR matrix with spectra represented in rows.}

\item{ppm}{ppm vector.}

\item{shift}{Chemical shift region to be plotted (in ppm).}

\item{interactive}{logical, interactive version (plotly package)}

\item{...}{Additional parameters to be passed on to the graphics generic plot function.}
}
\value{
Plot showing NMR spectra
}
\description{
Simple plotting of multiple NMR spectra overlayed
}
\details{
Low-level plotting function for NMR spectra, interactive plotting with ggplotly
}
\section{}{
NA
}

\examples{
data(covid)
matspec(X[1:2,], ppm) # interactive
matspec(X[1:2,], ppm, interactive=FALSE) # static
}
\seealso{
\code{\link{spec}} \code{\link{plot}}

Other NMR: 
\code{\link{alignSegment}()},
\code{\link{binning}()},
\code{\link{bline}()},
\code{\link{excise1d}()},
\code{\link{get.idx}()},
\code{\link{lw}()},
\code{\link{noise.est}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{spec}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{stocsy}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR}
