% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllViz.R
\name{matspec}
\alias{matspec}
\title{Simple plotting of multiple NMR spectra overlayed}
\usage{
matspec(X, ppm, shift = c(0, 9.5), interactive = T, ...)
}
\arguments{
\item{X}{NMR matrix with spectra represented in rows.}

\item{ppm}{ppm vector.}

\item{shift}{Chemical shift region to be plotted (in ppm).}

\item{interactive}{logical, interactive version (plotly package)}

\item{...}{Additional parameters to be passed on to the graphics generic plot function.}
}
\description{
Simple plotting of multiple NMR spectra overlayed
}
\details{
Low-level plotting function for NMR spectra, interactive plotting with ggplotly
}
\section{}{
NA
}

\seealso{
\code{\link{spec}} \code{\link{plot}}

Other visNMR: 
\code{\link{spec}()}
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
\concept{visNMR}
