% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Perform Principal Component Analysis}
\usage{
pca(X, pc = 2, scale = "UV", center = T, method = "nipals")
}
\arguments{
\item{X}{Numeric input matrix with each row representing an observation and each column a metabolic feature.}

\item{pc}{Desired number of principal components.}

\item{scale}{Desired scaling method: \code{None}, \code{UV} (unit variance) or \code{Pareto} (Pareto scaling).}

\item{center}{Logical indicating if data should be mean centered.}

\item{method}{Algorithm for computing PCA. NIPALS is standard and usually fine, see Details for other methods.}
}
\value{
This function returns a \emph{PCA_MetaboMate} S4 object.
}
\description{
This function is used to perform Principal Component Analysis (PCA).
}
\details{
Other PCA methods: These rely on implementation inf the pcaMethods R package and include 'svd', 'ppca', 'svdImpute', 'robustPca', 'nlpca'. For complete list of available methods see pca function documentation of the \code{pcaMethods} package.
}
\references{
Geladi, P and Kowalski, B.R. (1986), Partial least squares and regression: a tutorial. \emph{Analytica Chimica Acta}, 185, 1-17.
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
\concept{dataviz nmr functions, ms functions}
