% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PCA_metabom8-class}
\alias{PCA_metabom8-class}
\title{An S4 class to represent an PCA model constructed with metabom8}
\description{
An S4 class to represent an PCA model constructed with metabom8
}
\section{}{
NA
}

\seealso{
Other NMR & MS: 
\code{\link{OPLS_metabom8-class}}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR & MS}
