% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotload_cat.R
\name{plotload_cat}
\alias{plotload_cat}
\title{Plot PCA or O-PLS loadings}
\usage{
plotload_cat(obj, pc, an, title = NULL, legend = T, ...)
}
\arguments{
\item{obj}{PCA or OPLS object from packages metabom8 or MetaboMate}

\item{pc, }{num vector descrbing which components to visualise in on abcissa and ordinate (in case of OPLS the first element in pc is ignored as abcissa always shows predicted component)}

\item{an}{list, three elements describing colour, point shape and point labels (all is aes context)}

\item{title}{Plot title}

\item{legend}{logical, indicating if legend should be included (legend takes up too much space when high number of categorical varaibles)}

\item{...}{Optional arguments passed to \code{geom_point()}}
}
\value{
data.frame containing H.T2 ellipse cooredinates
}
\description{
Visualisation of loadings using scatter plot
}
\section{}{
NA
}

\references{
Geladi, P and Kowalski, B.R. (1986), Partial least squares and regression: a tutorial. \emph{Analytica Chimica Acta}, 185, 1-17.
}
\seealso{
Other dataviz: 
\code{\link{plotload}()},
\code{\link{plotscores}()},
\code{\link{specload}()}
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}

Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
\concept{dataviz}
