% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotscores.R
\name{plotscores}
\alias{plotscores}
\title{Plotting PCA or OPLS model scores}
\usage{
plotscores(obj, pc, an, title = "", qc, legend = "in", cv = T, ...)
}
\arguments{
\item{obj}{PCA, PLS or OPLS model of the package \emph{metabom8} or \emph{MetaboMate}.}

\item{pc}{num vector (n=2), indicating which omponents to be plotted on abscissa and ordinate. If not specified: PCA components 1 vs 2, OPLS: predictive vs 1st orth component.}

\item{an}{list of 3 elements: 1st: scatter colour, 2nd: shape and 3rd: label specification (see Details).}

\item{title}{cher, plot title.}

\item{qc}{int vector, row indices of QC samples (can be left NA). If specified, QC samples will be highlighted in the plot.}

\item{legend}{str, position of the plot legend, set to NA if legend should be outside of the plotting area.}

\item{cv}{logical, indicating if cross-validated (cv) scores should be plotted (only for PLS and O-PLS).}

\item{...}{Additional values passed on to \code{\link[ggplot2]{scale_colour_gradientn}}.}
}
\value{
This function returns a \emph{ggplot2} plot object.
}
\description{
Function for plotting PCA, PLS or OPLS model scores.
}
\details{
Scores colouring is specified with the argument \code{an}, which is a list of three elements. The first list element specifies the colour (class factor required for categorical variables), the second list element specifies a point labeling (class character or factor) and the third list element specifies point shape. The Hotelling's \out{T<sup>2</sup>} ellipse is automatically included and calculated for the dimensions selected by the \code{pc} argument.
}
\section{}{
NA
}

\references{
Trygg J. and Wold, S. (2002) Orthogonal projections to latent structures (O-PLS). \emph{Journal of Chemometrics}, 16.3, 119-128.

Hotelling, H. (1931) The generalization of Student’s ratio. \emph{Ann. Math. Stat.}, 2, 360-378.
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}

Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
