% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocsy.R
\name{plotStocsy}
\alias{plotStocsy}
\title{Plot Statistical Total Correlation Spectroscopy (STOCSY)}
\usage{
plotStocsy(stoc_mod, shift = c(0, 10), title = NULL)
}
\arguments{
\item{stoc_mod}{STOCSY object, as created with function 'stocsy' ('see ?stocsy()')}

\item{shift}{num array, chemical shift range defining plotting area}

\item{title}{char, plot title}
}
\description{
Plot Statistical Total Correlation Spectroscopy (STOCSY)
}
\section{}{
NA
}

\seealso{
Other metID functions: 
\code{\link{stocsy}()}
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR functions}
\concept{metID functions}
