% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocsy.R
\name{plotStocsy}
\alias{plotStocsy}
\title{Plot Statistical Total Correlation Spectroscopy (STOCSY)}
\usage{
plotStocsy(stoc_mod, shift = c(0, 10), title = NULL)
}
\arguments{
\item{stoc_mod}{STOCSY object, as created with function 'stocsy' ('see ?stocsy()')}

\item{shift}{num array, chemical shift range defining plotting area}

\item{title}{char, plot title}
}
\value{
ggplot2 object, stocsy plot
}
\description{
Plot Statistical Total Correlation Spectroscopy (STOCSY)
}
\section{}{
NA
}

\examples{
data(covid)
stcy_glucose=stocsy(X, ppm, driver=5.233)
plotStocsy(stcy_glucose, shift=c(5.15, 5.30), title='Alpha-anomeric proton of glucose (doublet)')
points(speaks[[1]]$ppm, speaks[[1]]$Int, col=factor(speaks[[1]]$type))
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR functions}
