% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocsy.R
\name{plotStocsy}
\alias{plotStocsy}
\title{Plot Statistical Total Correlation Spectroscopy (STOCSY)}
\usage{
plotStocsy(stoc_mod, shift = c(0, 10), title = NULL)
}
\arguments{
\item{stoc_mod}{An object of class \code{stocsy1d_metabom8} returned by the \code{stocsy()} function.}

\item{shift}{Numeric vector of length 2 specifying the chemical shift range (ppm) to be plotted.}

\item{title}{Character string for the plot title.}
}
\value{
A \code{ggplot2} object displaying the STOCSY spectrum within the specified shift range.
}
\description{
Generates a plot of the results from a STOCSY analysis.
}
\examples{
data(covid)
X <- covid$X
ppm <- covid$ppm
sy_glucose <- stocsy(X, ppm, driver = 5.233)
plotStocsy(sy_glucose, shift = c(5.15, 5.30), title = "Alpha-anomeric proton of glucose (doublet)")

}
