% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.r
\name{pls}
\alias{pls}
\title{Partial Least Squares (PLS)}
\usage{
pls(
  X,
  Y,
  center = TRUE,
  scale = "UV",
  cv = list(method = "k-fold_stratified", k = 7, split = 2/3),
  maxPCo = 5,
  plotting = TRUE
)
}
\arguments{
\item{X}{A numeric matrix or data frame. Each row represents an observation, and each column a metabolic variable.}

\item{Y}{Response vector or matrix. Must match the number of rows in \code{X}.}

\item{center}{Logical. Should data be mean-centered? Default is \code{TRUE}.}

\item{scale}{Character. Scaling method: \code{"None"}, \code{"UV"} (unit variance), or \code{"Pareto"}.}

\item{cv}{Named list specifying cross-validation settings:
\describe{
\item{\code{method}}{Cross-validation type: \code{"k-fold"}, \code{"k-fold_stratified"}, \code{"MC"}, or \code{"MC_balanced"}.}
\item{\code{split}}{Fraction of observations used for training (used in Monte Carlo CV).}
\item{\code{k}}{Number of folds or repetitions.}
}}

\item{maxPCo}{Integer. Maximum number of orthogonal components to test.}

\item{plotting}{Logical. If \code{TRUE}, model summary (e.g. R2X, Q2, AUROC) is plotted. Default is \code{TRUE}.}
}
\value{
An object of class \code{PLS_metabom8}, an S4 class with scores, loadings, predictions, and validation statistics.
}
\description{
Fit a Partial Least Squares (PLS) model for regression (R) or discriminant analysis (DA), with automated cross-validated component selection.
}
\details{
Cross-validation is used to select the optimal number of predictive components based on Q2 or AUROC. The method supports both regression and classification with binary or multi-class responses. Model interpretability is often best with pairwise class comparisons.
}
\examples{
data(covid)
model <- pls(X, Y = an$type)
plotscores(model, an = list(Class = an$type, Clinic = an$hospital, id = 1:nrow(an)), pc = c(1, 2))

}
\references{
Geladi, P. & Kowalski, B.R. (1986). Partial least squares regression: a tutorial. \emph{Analytica Chimica Acta}, 185, 1–17.
}
