% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppick.R
\name{ppick}
\alias{ppick}
\title{Peak Picking}
\usage{
ppick(X, ppm, fil_p = 3, fil_n = 5)
}
\arguments{
\item{X}{num matrix, NMR matrix with spectra in rows}

\item{ppm}{num vec, chemical shift matching to X}

\item{fil_p, }{filter order}

\item{fil_n}{num, filter length (must be odd)}
}
\value{
Peak picked
}
\description{
Peak Picking
}
\details{
Function smoothes the spectrum to de-emphasise the influence of noise and then identifies peak according to change of sign in adjacent points. Smoothing is performed with a Savitzki-Golay filter.
}
\seealso{
\code{\link[signal]{sgolayfilt}}

Other NMR: 
\code{\link{binning}()},
\code{\link{bline}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()}
}
\author{
Torben Kimhofer \email{torben.kimhofer@gmail.com}
}
\concept{NMR}
