% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_opls.R
\name{predict_opls}
\alias{predict_opls}
\alias{pred.opls}
\title{OPLS model predictions}
\usage{
predict_opls(opls_model, newdata, idx_scale = NULL)
}
\arguments{
\item{opls_model}{OPLS model (regression of discriminant analysis) of class \code{OPLS_MetaboMate}.}

\item{newdata}{NMR data matrix or dataframe with rows representing spectra and identical features in columns as data matrix used to calculate original OPLS model.}

\item{idx_scale}{int vector, row-indices of newdata used to subselect samples to determine scale and center pars. Recommded: set to NULL: use center and scaling parameters from opsl training data}
}
\value{
Returned is a list with the following elements:
\describe{
\item{Y_predicted}{Class or numeric outcome predictions for discriminant analysis or regression, repspectively.}
\item{t_pred}{Predicted OPLS model scores for predictive component(s).}
\item{t_orth}{Predicted OPLS model scores for orthogonal component(s).}
\item{t_orth_pca}{Scores of a PCA model (first component) calculated using all predicted OPLS orthogonal component scores - only done when there are more than one orthogonal components in \code{opls_model}.}
}
}
\description{
Calculation of OPLS model predictions using new data
}
\details{
Class predictions for discriminant analysis are not adjusted for unbalanced sample sizes and therefore, predictions can be biased towards the group with the largest number of samples. The list element \code{t_orth_pca} represent scores of the first principal component of a PCA model caclulated with all orthogonal components, therefore, summarises all orthogonal components into a single one. This can only be done if there are more than one orthogonal components in \code{opls_modelel}, otherwise this list element is \code{NULL}.
}
\references{
Trygg J. and Wold, S. (2002) Orthogonal projections to latent structures (O-PLS). \emph{Journal of Chemometrics}, 16.3, 119-128.

Geladi, P and Kowalski, B.R. (1986), Partial least squares and regression: a tutorial. \emph{Analytica Chimica Acta}, 185, 1-17.
}
\seealso{
\code{\link{opls}}
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
