% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1d.R
\name{read1d}
\alias{read1d}
\title{Read-in 1D NMR spectra}
\usage{
read1d(path, procs_exp = 1, n_max = 1000, filter = T)
}
\arguments{
\item{path}{char, File directory containing spectra}

\item{procs_exp}{num, Topspin processing experiment ID}

\item{n_max}{int, Maximum number of spectra to read-in}

\item{filter}{lobic, filter for intact file systems (TRUE is recommended)}
}
\description{
Read-in 1D NMR spectra
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
