% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1d.R
\name{read1d}
\alias{read1d}
\alias{read1d_proc}
\title{Import 1D NMR spectra (TopSpin processed)}
\usage{
read1d(
  path,
  exp_type = list(pulprog = "noesygppr1d"),
  n_max = 1000,
  filter = TRUE,
  recursive = TRUE,
  verbose = 1
)

read1d_proc(
  path,
  exp_type = list(pulprog = "noesygppr1d"),
  n_max = 1000,
  filter = TRUE,
  recursive = TRUE,
  verbose = 1
)
}
\arguments{
\item{path}{Character. Directory path containing Bruker NMR experiments.}

\item{exp_type}{Named list. Filter experiments by acquisition parameters (e.g., list(pulprog = 'noesygppr1d')).}

\item{n_max}{Integer. Maximum number of spectra to import. Default: 1000.}

\item{filter}{Logical. Filter out experiments with incomplete file systems.}

\item{recursive}{Logical. Search path recursively. Default: TRUE.}

\item{verbose}{Logical or numeric. Verbosity level.}
}
\value{
The following objects are assigned to the global environment:
\itemize{
\item \code{X}: Numeric matrix of spectra (samples x variables)
\item \code{ppm}: Numeric vector of chemical shift values
\item \code{meta}: Data frame of acquisition/processing metadata
}
Also returns a (named) list invisibly for testing.
}
\description{
Imports Bruker TopSpin-processed 1D NMR spectra. Spectrometer and processing
parameters are extracted from \code{acqus} and \code{procs} files. Optionally filters
experiments using acquisition metadata.
}
\examples{
path <- system.file("extdata", package = "metabom8")
read1d(path, exp_type = list(pulprog = "noesygppr1d"), n_max = 2)

}
\seealso{
Other NMR: 
\code{\link{alignSegment}()},
\code{\link{binning}()},
\code{\link{cvanova}()},
\code{\link{get_idx}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{normErectic}()},
\code{\link{read1d_raw}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{storm}()}
}
\concept{NMR}
