% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1d.R
\name{read1d}
\alias{read1d}
\alias{read1d_proc}
\title{Import 1D NMR spectra (TopSpin processed)}
\usage{
read1d(
  path,
  exp_type = list(pulprog = "noesygppr1d"),
  n_max = 1000,
  filter = TRUE,
  recursive = TRUE,
  verbose = 1,
  to_global = TRUE
)

read1d_proc(
  path,
  exp_type = list(pulprog = "noesygppr1d"),
  n_max = 1000,
  filter = TRUE,
  recursive = TRUE,
  verbose = 1,
  to_global = TRUE
)
}
\arguments{
\item{path}{Character. Directory path containing Bruker NMR experiments.}

\item{exp_type}{Named list. Filter experiments by acquisition parameters (e.g., list(pulprog = 'noesygppr1d')).}

\item{n_max}{Integer. Maximum number of spectra to import. Default: 1000.}

\item{filter}{Logical. Filter out experiments with incomplete file systems.}

\item{recursive}{Logical. Search path recursively. Default: TRUE.}

\item{verbose}{Logical or numeric. Verbosity level.}

\item{to_global}{Logical. Objects will be added to global environment}
}
\value{
A named list with three elements:
\describe{
\item{X}{A numeric matrix of spectra (rows = samples, columns = ppm values).}
\item{ppm}{A numeric vector of chemical-shift axis (ppm).}
\item{meta}{A data frame of acquisition/processing metadata, row-aligned with \code{X}.}
}

If \code{to_global = TRUE}, these objects are also assigned to the global environment.
In that case, any existing objects with the same names will be overwritten.
}
\description{
Imports TopSpin-processed 1D NMR spectra, together with spectrometer acquisition
and TopSpin processing parameters (\code{acqus} and \code{procs}, respectively).
}
\examples{
path <- system.file("extdata", package = "metabom8")
read1d(path, exp_type = list(pulprog = "noesygppr1d"), n_max = 2)

}
