% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1d_raw.R
\name{read1d_raw}
\alias{read1d_raw}
\title{Read-in 1D NMR spectra}
\usage{
read1d_raw(
  path,
  n_max = 1000,
  filter = T,
  apodisation = list(fun = "exponential", lb = 0.2),
  zerofil = 1L,
  return = "absorption",
  pulprog = "<noesygppr1d>",
  verbose = TRUE,
  recursive = TRUE
)
}
\arguments{
\item{path}{char, File directory containing spectra}

\item{n_max}{int, Maximum number of spectra to read-in}

\item{filter}{lobic, filter for intact file systems (TRUE is recommended)}

\item{apodisation}{list, apodisation type and parameter}

\item{zerofil}{int, zero-filling, exponent summand of base 2}

\item{return}{char, return mode: absorption, dispersion, magnitude}

\item{pulprog}{char, experiment type to read-in as defined by TopSpin pulprog (e.g. <noesygppr1d>)}
}
\value{
Three objects: NMR data matrix (2D: rows=spectra, cols=chem shift variables), ppm num vector matched to NMR data columns, meta data.frame containing spectrometer metadata
}
\description{
Read-in 1D NMR spectra
}
\section{}{
NA
}

\seealso{
\code{\reference{read1d}}
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
\concept{Import NMR data functions}
