% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1d_raw.R
\name{read1d_raw}
\alias{read1d_raw}
\title{Read-in FIDs and process to spectra}
\usage{
read1d_raw(
  path,
  exp_type = list(exp = c("PROF_PLASMA_CPMG128_3mm"), pulprog = c("noesygppr1d")),
  apodisation = list(fun = "exponential", lb = 0.2),
  zerofil = 1L,
  return = "absorption",
  verbose = 1,
  recursive = TRUE,
  n_max = 1000,
  filter = TRUE
)
}
\arguments{
\item{path}{char, path to file directory containing spectra}

\item{exp_type}{named list, filter for acquisition paramters of experiments to read-in (see Details)}

\item{apodisation}{named list, apodisation function and its parameters (see Details)}

\item{zerofil}{int, amount of zeros to append to FID given as exponent added of base 2 (see Details)}

\item{return}{char, return mode of spectra: absorption, dispersion or magnitude mode}

\item{verbose}{num, different verbose levels: 0 (no info), 1 (overview), 2 (detailed), 3 (step-by-step for debugging)}

\item{recursive}{logic, if TRUE recursively search all subfolders of path for specified NMR files}

\item{n_max}{int, maximum number of experiments to read-in}

\item{filter}{logic, remove experiments with incomplete file systems (TRUE is recommended)}
}
\value{
The function exports the following three objects into the currently active R environment (no variable assignments needed):
\itemize{
\item X, num matrix:  NMR data, spectra in rows
\item ppm, num array - chemical shift positions, length matches to columns in X
\item meta, data.frame - spectrometer metadata as extracted from individual \emph{acqus} files, row-matched to X
}
Objects in the R environment with the same variable names will be overwritten.
}
\description{
Read-in FIDs and process to spectra
}
\details{
In the first step, read-in are FIDs generated with experimental condition(s) specified with the exp_type argument. This represents a list with each element representing a parameter condition, named according to spectrometer parameters listed in \emph{acqus} file. For example, to read standard 1D NMR experiments use \code{exp_type=list(exp='noesygppr1d')}. More than one argument can be provided as list element.

The apodisation argument is a named list specifying the function name in element \emph{fun} and functions-specific paramter arguments. There are the following different apodisation functions and arguments:
#' @return
\itemize{
\item exponential, arguments: lb (line broadening factor)
\item cosine, no further arguments
\item sine, no further arguments
\item sem, combined sine-bell - exponential fct: arguments: lb (line broadening factor)
}

The zerofil argument specifies the amount of zeros to append to the FID and is expressed as exponent addand in the binary numeral system: \code{2^(1+x)}, with x being the zerofil parameter argument. Hence, \code{zerofil=1} doubles the amount of data points.
}
\section{}{
NA
}

\examples{
path<-system.file('extdata/',  package = 'metabom8')
read1d_raw(path,  exp_type=list(exp='PROF_PLASMA_NOESY'), apodisation=list(fun='exponential', lb=0.2), n_max=3)
}
\seealso{
\code{\link[=read1d]{Import TopSpin processed spectra}}

Other NMR: 
\code{\link{alignSegment}()},
\code{\link{binning}()},
\code{\link{bline}()},
\code{\link{get.idx}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{spec}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{stocsy}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR}
