% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllViz.R
\name{spec}
\alias{spec}
\title{Simple plotting of a single NMR spectrum}
\usage{
spec(x, ppm, shift = c(0, 11), add = F, interactive = T, name = "A", ...)
}
\arguments{
\item{x}{NMR spectrum.}

\item{ppm}{ppm vector.}

\item{shift}{chemical shift region to be plotted.}

\item{add}{Logical indicating if spectrum should be added to a current plot generated with \code{spec()} or \code{matspec()}.}

\item{interactive}{logical, interactive version (plotly package)}

\item{...}{Additional parameters to be passed on to the graphics generic plot function.}
}
\description{
Simple plotting of a single NMR spectrum
}
\details{
Low-level plotting function for a single NMR spectrum.
}
\section{}{
NA
}

\seealso{
\code{\link{matspec}} \code{\link{plot}}
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
