% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllViz.R
\name{spec}
\alias{spec}
\title{Simple plotting of a single NMR spectrum}
\usage{
spec(
  x,
  ppm,
  shift = c(0, 11),
  add = FALSE,
  interactive = TRUE,
  name = "A",
  mode = "lines",
  ...
)
}
\arguments{
\item{x}{NMR spectrum.}

\item{ppm}{ppm vector.}

\item{shift}{chemical shift region to be plotted.}

\item{add}{Logical indicating if spectrum should be added to a current plot generated with \code{spec()} or \code{matspec()}.}

\item{interactive}{logical, interactive version (plotly package)}

\item{name}{string, name of trace (only used in interactive mode)}

\item{mode}{string, plot mode for interactive version: 'lines', 'lines+markers' or 'markers' (see Details)}

\item{...}{Additional parameters to be passed on to the graphics generic plot function.}
}
\value{
Plot showing NMR spectrum
}
\description{
Simple plotting of a single NMR spectrum
}
\details{
Low-level plotting function for a single NMR spectrum (base graphics). Interactive visualisation in mode of markers or line+markers will take longer (mode='lines' is recommended).
}
\section{}{
NA
}

\examples{
data(covid)
spec(X[1,], ppm) # interactive
spec(X[2,], ppm, add=TRUE) # add trace
spec(X[1,], ppm, interactive=FALSE) # static
}
\seealso{
\code{\link{matspec}} \code{\link{plot}}

Other NMR: 
\code{\link{binning}()},
\code{\link{bline}()},
\code{\link{get.idx}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{stocsy}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR}
