% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllViz.R
\name{spec}
\alias{spec}
\title{Plot single NMR spectrum}
\usage{
spec(
  x,
  ppm,
  shift = c(0, 11),
  add = FALSE,
  interactive = TRUE,
  name = "A",
  mode = "lines",
  ...
)
}
\arguments{
\item{x}{num array, NMR spectrum.}

\item{ppm}{num array, ppm vector.}

\item{shift}{num array, chemical shift region to be plotted.}

\item{add}{logical, indicating if spectrum should be added to a current plot generated with \code{spec()} or \code{matspec()}.}

\item{interactive}{logical, interactive version (plotly package)}

\item{name}{string, name of trace (only used in interactive mode)}

\item{mode}{string, plot mode for interactive version: 'lines' (recommended), 'lines+markers' or 'markers'}

\item{...}{Additional parameters to be passed on to the graphics generic plot function.}
}
\description{
Plot single NMR spectrum
}
\details{
Non-interactive mode: Base graphics for single NMR spectrum. Interactive mode: Plotly graphics.
}
\section{}{
NA
}

\examples{
data(covid)
spec(X[1,], ppm) # interactive
spec(X[2,], ppm, add=TRUE) # add trace
spec(X[1,], ppm, interactive=FALSE) # static
}
\seealso{
\code{\link{matspec}} \code{\link{plot}}

Other NMR: 
\code{\link{alignSegment}()},
\code{\link{bline}()},
\code{\link{excise1d}()},
\code{\link{get.idx}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{normErectic}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{stocsy}()},
\code{\link{storm}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR}
