% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllViz.R
\name{specload}
\alias{specload}
\title{Overlay Loadings with NMR Spectra}
\usage{
specload(
  mod,
  shift = c(0, 10),
  an,
  alp = 0.3,
  size = 0.5,
  pc = 1,
  type = "Backscaled",
  title = "",
  r_scale = FALSE
)
}
\arguments{
\item{mod}{A model object of class \code{PCA_metabom8} or \code{OPLS_metabom8}.}

\item{shift}{Numeric vector (length 2), chemical shift region to visualize (in ppm).}

\item{an}{List of up to three grouping variables (for facet, color, linetype). First must be defined.}

\item{alp}{Numeric (0–1). Alpha level for spectra.}

\item{size}{Numeric. Line width for plotted spectra.}

\item{pc}{Integer. Component to visualize (for OPLS, set to 1).}

\item{type}{Character. Either \code{"Statistical reconstruction"} or \code{"Backscaled"} (case-insensitive).}

\item{title}{Character. Plot title.}

\item{r_scale}{Logical. If \code{TRUE}, correlation color scale is fixed to \code{[0, 1]}. Only used in statistical reconstruction.}
}
\value{
A \code{ggplot} object.
}
\description{
This function overlays loadings from a PCA or OPLS model on top of NMR spectra using \code{ggplot2}.
It supports two types of loadings visualization: statistical reconstruction or backscaling.
}
\details{
\itemize{
\item \strong{Statistical reconstruction}: Correlates predictive scores with spectral variables.
\item \strong{Backscaled}: Multiplies model loadings by feature SDs.
}
}
\examples{
data(covid)
X <- covid$X
an <- covid$an

model <- pca(X)
specload(model, shift = c(1, 2), an = list(an$type), pc = 1, alp = 0.8)

}
\seealso{
\code{\link{specOverlay}}, \code{\link{pca}}, \code{\link{opls}}
}
