% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocsy.R
\name{stocsy}
\alias{stocsy}
\title{Statistical Total Correlation Spectroscopy (STOCSY)}
\usage{
stocsy(X, ppm, driver, plotting = TRUE, title = NULL)
}
\arguments{
\item{X}{Numeric matrix or data frame where rows are spectra and columns are chemical shift variables.}

\item{ppm}{Numeric vector of chemical shift values corresponding to columns in \code{X}.}

\item{driver}{Numeric scalar or vector. A ppm value (for internal STOCSY) or external numeric vector of same length as number of rows in \code{X}.}

\item{plotting}{Logical. If TRUE, display a STOCSY plot.}

\item{title}{Character string for the plot title.}
}
\value{
An S4 object of class \code{stocsy1d_metabom8} containing correlation, covariance, driver, and metadata.
}
\description{
Performs STOCSY analysis on 1D NMR spectra. The function calculates correlation and covariance of all variables in \code{X} against a given driver variable (either a ppm value or external vector).
}
\examples{
data(covid)
stcy_glucose <- stocsy(X, ppm, driver = 5.233)
plotStocsy(stcy_glucose, shift = c(5.15, 5.30), title = "Alpha-anomeric proton of glucose (doublet)")

}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
