% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocsy.R
\name{stocsy}
\alias{stocsy}
\title{Statistical Total Correlation Spectroscopy (STOCSY)}
\usage{
stocsy(X, ppm, driver, plotting = TRUE, title = NULL)
}
\arguments{
\item{X}{Numeric matrix or dataframe where each row represents a NMR spectrum and each column a chemical shift variable.}

\item{ppm}{num array of chemical shift variables, matched to columns in X}

\item{driver}{num, chemical shift indicating STOCSY driver signal}

\item{plotting}{logical, indicing if results should be plotted}

\item{title}{char, plot title}
}
\value{
S4 object of class stocsy1d_metabom8
}
\description{
Statistical Total Correlation Spectroscopy (STOCSY)
}
\section{}{
NA
}

\examples{
data(covid)
stcy_glucose=stocsy(X, ppm, driver=5.233)
plotStocsy(stcy_glucose, shift=c(5.15, 5.30), title='Alpha-anomeric proton of glucose (doublet)')
points(speaks[[1]]$ppm, speaks[[1]]$Int, col=factor(speaks[[1]]$type))
}
\seealso{
Other NMR: 
\code{\link{alignSegment}()},
\code{\link{bline}()},
\code{\link{excise1d}()},
\code{\link{get.idx}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{normErectic}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{spec}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{storm}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR}
