% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocsy.R
\name{stocsy}
\alias{stocsy}
\title{Statistical Total Correlation Spectroscopy (STOCSY)}
\usage{
stocsy(X, ppm, driver, plotting = T, title = NULL, pltly = T)
}
\arguments{
\item{X}{Numeric matrix or dataframe where each row represents a NMR spectrum and each column a chemical shift variable.}

\item{ppm}{num array of chemical shift variables, matched to columns in X}

\item{driver}{num, chemical shift indicating STOCSY driver signal}

\item{plotting}{logical, indicing if results should be plotted}

\item{title}{char, plot title}
}
\description{
Statistical Total Correlation Spectroscopy (STOCSY)
}
\section{}{
NA
}

\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
