% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{stocsy1d_metabom8-class}
\alias{stocsy1d_metabom8-class}
\title{STOCSY model object from metabom8}
\description{
An S4 class to represent a STOCSY model.
}
\section{Slots}{

\describe{
\item{\code{version}}{Character string indicating version of the model}

\item{\code{X}}{Spectral data matrix}

\item{\code{ppm}}{Chemical shift values (vector)}

\item{\code{driver}}{Numeric vector indicating driver variable}

\item{\code{r}}{Correlation coefficients (vector)}

\item{\code{cov}}{Covariance values (vector)}
}}

\seealso{
Other NMR: 
\code{\link{alignSegment}()},
\code{\link{binning}()},
\code{\link{cvanova}()},
\code{\link{get_idx}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{normErectic}()},
\code{\link{read1d_raw}()},
\code{\link{storm}()}
}
\author{
\email{tkimhofer@gmail.com}
}
\concept{NMR}
