% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{stocsy1d_metabom8-class}
\alias{stocsy1d_metabom8-class}
\title{An S4 class to represent STOCSY model constructed with metabom8}
\description{
An S4 class to represent STOCSY model constructed with metabom8
}
\section{}{
NA
}

\seealso{
Other NMR: 
\code{\link{binning}()},
\code{\link{bline}()},
\code{\link{get.idx}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{spec}()},
\code{\link{stocsy}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}
}
\concept{NMR}
