% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm.R
\name{storm}
\alias{storm}
\title{Subset Optimisation by Reference Matching (STORM)}
\usage{
storm(
  X = bbin[[1]],
  ppm = bbin[[2]],
  b = 30,
  q = 0.05,
  idx.refSpec = NULL,
  shift = c(1.117, 1.25)
)
}
\arguments{
\item{X}{Numeric matrix or dataframe where each row represents a NMR spectrum and each column a chemical shift variable.}

\item{ppm}{num array of chemical shift variables, matched to columns in X}

\item{b}{int, boundary value provided as positive integer in chemical shift space}

\item{q}{num, level of significance}

\item{idx.refSpec}{int, reference spectrum defined as row index for X}

\item{shift}{num array, min and max chemical shift value defining narrow boundary of the target signal}
}
\value{
Vector of row indices that define the optimal subset.
}
\description{
Subset Optimisation by Reference Matching (STORM)
}
\details{
The STORM algorithm can be used to subselect NMR spectra based on a spectrocopic target signal to match multiplicity and chemical shift position. The output are row-indices for X, defining the optimal spectral subset that is typically used as input for STOCSY.
}
\references{
Posma, Joram M., et al. "Subset optimization by reference matching (STORM): an optimized statistical approach for recovery of metabolic biomarker structural information from 1H NMR spectra of biofluids." Analytical chemistry 84.24 (2012): 10694-10701.
}
\seealso{
stocsy

Other NMR: 
\code{\link{alignSegment}()},
\code{\link{bline}()},
\code{\link{excise1d}()},
\code{\link{get.idx}()},
\code{\link{lw}()},
\code{\link{matspec}()},
\code{\link{noise.est}()},
\code{\link{normErectic}()},
\code{\link{plotload}()},
\code{\link{ppick}()},
\code{\link{read1d_raw}()},
\code{\link{read1d}()},
\code{\link{specload}()},
\code{\link{spec}()},
\code{\link{stocsy1d_metabom8-class}},
\code{\link{stocsy}()}
}
\author{
\email{torben.kimhofer@murdoch.edu.au}

torben.kimhofer@murdoch.edu.au
}
\concept{NMR}
